### 2022-08-22 #4695 @editorialbot
[PRE REVIEW]: riversCentralAsia: An R package for hydrological modelling
**Submitting author:** @mabesa (<a href="http://orcid.org/0000-0003-2089-3478">Beatrice Marti</a>)
**Repository:** https://github.com/hydrosolutions/riversCentralAsia/
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.1.0
