### 2022-08-27 #4695 @crvernon
:wave: - @mengqi-z Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
