### 2022-09-27 #4805 @editorialbot
[REVIEW]: riversCentralAsia: An R package to support data pre- and postprocessing for hydrological modelling with RS MINERVE
**Submitting author:** @mabesa (<a href="http://orcid.org/0000-0003-2089-3478">Beatrice Marti</a>)
**Repository:** https://github.com/hydrosolutions/riversCentralAsia/
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.2.0
