### 2022-09-29 hydrosolutions/riversCentralAsia#97 @tonyewong
open source license?
For JOSS, there is a requirement that the software be shared under an OSI-approved open source license:

> License: Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?

At the moment, I don't see one (at least not at the top-level or in a license-specific directory). Can one be added?
