### 2022-09-29 #4805 @tonyewong
I see that EditorialBot found this as well, but I opened an issue in the repo, regarding an open source license (https://github.com/hydrosolutions/riversCentralAsia/issues/97). Can one be added to the top level of the repo and a short section in the top level README.md that summarizes license info?
