### 2022-09-29 #4805 @mabesa
Thank you for pointing that out. I added a license file and added a section in the README.
