### 2022-10-07 cbe2ca3
Update gen_HRU_Climate_CSV_RSMinerve.R
exactextractr::exact_extract does not yield correct output. Added a comparison with raster::extract results and an automated method switch in case the two methods produce inconsistent results (in which case the raster package is used).
