### 2022-10-10 #4805 @tonyewong
I don't know if it needs addressing, but just FYI - as I installed, the dependency on package "exactextractr" threw an error message. I had to install it separately in a first step, and say "no" when the installer asks "Do you want to install from sources the package which needs compilation? (Yes/no/cancel)". But once I did that, things installed and the example in the top-level README runs just fine.
