### 2022-10-10 #4805 @tonyewong
@mabesa  The example in the README works nicely out-of-box, but I'm having issues running the tests in [riversCentralAsia/tests/testthat.R](https://github.com/hydrosolutions/riversCentralAsia)/[tests](https://github.com/hydrosolutions/riversCentralAsia/tree/master/tests/testthat.R). Specifically, I get:
```
Error in `test_dir()`:
! No test files found
Run `rlang::last_error()` to see where the error occurred.
```
Can some additional guidance on running those tests be provided in the README? (or elsewhere)
