### 2022-10-11 #4805 @mabesa
@tonyewong
I also had problems with exactextractr very recently, it started returning wrong results after the last update. I think I should probably immediately have this documented here... I'm sorry, this is the first time I run through a software review. Thank you for noting the installation problem, I hadn't seen that one yet! I will try to solve this issue and let you know asap.

I devtools::test() runs without error in the main directory. Also, I can navigate to ./tests and run testthat.R without problems or run tests of individual functions without error messages. Could you please specify how you ran the tests so I can reproduce the error message?
