### 2022-10-12 40ed2ee
tests and check(vignette = FALSE) successful
