### 2022-10-12 #4805 @crvernon
📣 Mid-week rally!

👋 @tonyewong  - It looks like you and @mabesa are working a few things out but all is going well.  Let me know if you need anything.

👋 @mengqi-z  - Can you provide a status update to how things are going on your side of things?  Thanks!

Keep up the good work!
