### 2022-10-12 #4805 @mabesa
Yep, I updated the package and tested it with the devtools utils and I am now re-testing the package with real-life data sets not included in the test suite. This procedure should be done tomorrow morning by the latest. I'll write again then, hopefully with an "issue fixed" happy-face emoji.
