### 2022-10-12 #4805 @mengqi-z
@crvernon - I am planning to get more reviews done by the end of this week.

@mabesa - a quick note for the installation step. The built-in pip operator "|>" used in the package requires R version >= 4.1. Could you add this requirement to the installation instruction? Thanks.
