### 2022-10-13 #4805 @tonyewong
@mabesa Looks like the more detailed error I see both when running `devtools::test()` or `source("testthat.R")` from the `tests` directory is the following:
```
══ Failed tests ══════════════════════════════════════════════════════════════════════
── Error (test_gen_basinElevationBands.R:9): Elevation bands are correctly generated ──
Error in `Ops.units(size_fxn(singles), threshold)`: both operands of the expression should be "units" objects
Backtrace:
    ▆
 1. └─riversCentralAsia::gen_basinElevationBands(...) at test_gen_basinElevationBands.R:9:2
 2.   ├─smoothr::drop_crumbs(...) at riversCentralAsia/R/gen_basinElevationBands.R:60:2
 3.   └─smoothr:::drop_crumbs.sf(...)
 4.     ├─smoothr::drop_crumbs(...)
 5.     └─smoothr:::drop_crumbs.sfc(...)
 6.       ├─singles[size_fxn(singles) >= threshold]
 7.       ├─sf:::`[.sfc`(singles, size_fxn(singles) >= threshold)
 8.       └─units:::Ops.units(size_fxn(singles), threshold)
── Error (test_gen_basinElevationBands.R:21): Can read dem from raster object ──
Error in `Ops.units(size_fxn(singles), threshold)`: both operands of the expression should be "units" objects
Backtrace:
    ▆
 1. └─riversCentralAsia::gen_basinElevationBands(...) at test_gen_basinElevationBands.R:21:2
 2.   ├─smoothr::drop_crumbs(...) at riversCentralAsia/R/gen_basinElevationBands.R:60:2
 3.   └─smoothr:::drop_crumbs.sf(...)
 4.     ├─smoothr::drop_crumbs(...)
 5.     └─smoothr:::drop_crumbs.sfc(...)
 6.       ├─singles[size_fxn(singles) >= threshold]
 7.       ├─sf:::`[.sfc`(singles, size_fxn(singles) >= threshold)
 8.       └─units:::Ops.units(size_fxn(singles), threshold)

[ FAIL 2 | WARN 0 | SKIP 1 | PASS 78 ]
```

Relatedly, I'd suggest just adding a line or two to the README.md to indicate how users ought to test their installs. I got the previous error by running some of the contents of the `testthat.R` script interactively. Pointing folks straight to `devtools::test()` would be an easy way to go.
