### 2022-10-13 2a066ca
Update _pkgdown.yml
Removed description for reference entry Package description
