+++ _pkgdown.yml
@@ -3,7 +3,7 @@ template:
   bootswatch: litera
 reference:
 - title: Package description
-  desc: Description of the package
+  desc:  
 - contents:
   - has_concept("Package description")
 - title: Pre-processing
