### 2022-10-13 3fb17e5
debugging error in test
Uninstalled all dependencies and reinstalled package to reproduce tonyewong's error messages in issue #100
Raster no longer recognizes the crs format epsg:xxxxx so changed functions gen_basinElevationBands and gen_HRU_Climate_CSV_RSMinerve to read in the proj format.

TODO: Will need to update the vignettes and also the caham book!
