### 2022-10-13 b5cbd11
Added description of how to test the package on the system in README
hopefully understandable
