### 2022-10-13 ee2f5a4
Checked if all articles knit
also ran devtools::document(roclets = c('rd', 'collate', 'namespace', 'vignette'))
