### 2022-10-13 1827090
Delete 04-glacier-functions.md
