### 2022-10-13 5d5bdb8
Delete 05-snow-calibration.md
