+++ vignettes/05-snow-calibration.md
@@ -1,391 +1,0 @@
----
-title: "Manual calibration of snow water equivalent"
-author: "Beatrice Marti"
-date: '2022-10-13'
-bibliography: references.bib
-output: rmarkdown::html_vignette
-fig_caption: yes
-vignette: >
-  %\VignetteIndexEntry{Manual calibration of snow water equivalent}
-  %\VignetteEngine{knitr::knitr}
-  %\VignetteEncoding{UTF-8}
----
-
-
-
-
-```r
-library(tidyverse)
-library(lubridate)
-library(sf)
-
-devtools::install_github("hydrosolutions/riversCentralAsia")
-library(riversCentralAsia)
-
-# The Atbashy glacier demo data is available from 
-# https://www.dropbox.com/sh/r0lqggc77ka0uxd/AAChuIyLHHFIfAdgxNKiU2dpa?dl=0
-# Download the data and adapt the path below.  
-data_path <- "../../atbashy_glacier_demo_data/"
-```
-
-This chapter of the vignette shows how to compare simulated snow water equivalent (SWE) to third party data. These code snipets can be used to iteratively adapt the parameters of the snow modules in the HBV models to better represent the SWE of the data product.
-
-In places where SWE observations are not available, a binary variable of SWE larger than a threshold per hydrological response unit (HRU) can be compared to snow covered area (a MODIS product) in each HRU (see for example @parajka_value_2008).
-
-## Calibrating snow water equivalent (SWE)
-
-At the time of writing, RS Minerve does not yet support the automatic calibration of SWE. However, snow melt is a major contribution to discharge in Central Asia. Measurements of SWE are only very rarely available. Thus, re-analysis products like the High Mountain Asia Snow Reanalysis product are valuable resources to validate hydrolgoical modelling efforts.
-
-As automated calibration of SWE is not yet supported in RS Minerve, we propose a work around in R to iteratively adapt the parameters of the HBV snow modules.
-
-### Extract observed SWE
-
-Actually observed SWE is not typically available for hydrological modeling in many catchments in Central Asia, we therefore propose to use existing model products like the High Mountain Snow Reanalysis Product [@liu_spatiotemporal_2021] which is available through [NSIDC](https://nsidc.org/data/HMA_SR_D/). Login to USGS Earth Data, select the files for download for your model area and download the data (for example using the python download script). We will refer to the SWE data set from HMASR as observed data.
-
-The downloaded data is subsequently pre-processed to extract the daily average SWE per HRU on the valid pixels of the data set (note that HMASR is not available on areas with permanent snow or ice cover). The following code sniped shows how this can be done.
-
-
-```r
-library(tmap)
-library(sf)
-library(raster)
-library(tidyverse)
-library(lubridate)
-
-devtools::install_github("hydrosolutions/riversCentralAsia")
-library(riversCentralAsia)
-
-# Path to the data directory downloaded from the download link provided above. 
-# Here the data is extracted to a folder called atbashy_glacier_demo_data
-data_path <- "../"
-
-# Function to concatenate and mask the HMASR product and extract SWE for each 
-# HRU in the model. 
[…]
