### 2022-10-14 7918dcb
Update paper.bib
removed https::/doi from reference
