+++ paper.bib
@@ -3,7 +3,7 @@
   author = {Siegfried, Tobias and Marti, Beatrice},
   publisher = {hydrosolutions},
   url = {https://hydrosolutions.github.io/caham_book/},
-  doi = {https://doi.org/10.5281/zenodo.6350042},
+  doi = {10.5281/zenodo.6350042},
   year = 2022
 }
 
