+++ README.Rmd
@@ -23,17 +23,19 @@ The package has been developed within the frame of a master level course on appl
 
 While the package has been developed for the Central Asia region, most of the functions are generic and can be used for modelling projects anywhere in the world. 
 
+The most important functionalities of the package as well as the raw data that can be processed with the package are described in the the articles of the [project documentation site](https://hydrosolutions.github.io/riversCentralAsia/){target="_blank"} but the examples in the book [Modeling of Hydrological Systems in Semi-Arid Central Asia](https://hydrosolutions.github.io/caham_book/){target="_blank"} demonstrate the full range of functions available and how to use them in a workflow. 
+
 ## Statement of need
 
 Data preparation comes before hydrological modelling and is actually one of the biggest work chunks in the modelling process. This package includes a number of helper functions that can be connected to efficient workflows that automatize the data preparation process for hydrological modelling. The functionality includes: 
 
-- Efficient processing of present and future climate forcing, including hydro-meterological data from Central Asia and down-scaling of ERA5 re-analysis data
+- Efficient processing of present and future climate forcing, including hydro-meterological data from Central Asia ([time series](https://hydrosolutions.github.io/riversCentralAsia/articles/01-discharge-processing-examples.html){target="_blank"} and re-analysis data) and down-scaling of ERA5 re-analysis data (a more advanced topic which is described in the [course book](https://hydrosolutions.github.io/caham_book/climate_data.html){target="_blank"})
 
-- The preparation of GIS layers for automated model generation
+- The [preparation of GIS layers for automated model generation](https://hydrosolutions.github.io/riversCentralAsia/articles/02-preparation-of-climate-forcing.html){target="_blank"} and chapter [Geospatial data](https://hydrosolutions.github.io/caham_book/geospatial_data.html){target="_blank"} in the course book
 
-- Simplified modelling of glacier dynamics
+- [Volume area scaling of glaciers](https://hydrosolutions.github.io/riversCentralAsia/articles/04-glacier-functions.html){target="_blank"}
 
-- Post-processing of simulation results, e.g. computation of flow duration curves
+- Post-processing of simulation results, e.g. [extraction and visualisation of snow water equivalent](https://hydrosolutions.github.io/riversCentralAsia/articles/05-snow-calibration.html){target="_blank"} or [computation of flow duration curves](https://hydrosolutions.github.io/riversCentralAsia/reference/computeAnnualFlowDurationCurve.html){target="_blank"} 
 
 - I/O interface with the hydrologic-hydraulic modelling software [RS Minerve](https://crealp.ch/rs-minerve/){target="_blank"} which can be accessed within R using the package [RSMinerveR](https://github.com/hydrosolutions/RSMinerveR){target="_blank"}.  
 
@@ -47,14 +49,7 @@ Currently, a relatively complete dataset of the Chirchik River Basin with decada
 The R package [RSMinverveR](https://github.com/hydrosolutions/RSMinerveR){target="_blank"} allows the running of the hydrologic-hydraulic modelling software [RS Minerve](https://crealp.ch/rs-minerve/){target="_blank"} directly from R without using the RS Minerve user interface. This functionality is for advanced R and RS Minerve users that wish to further speed up their modelling workflow.     
 
 ## Installation
-
-The package riversCentralAsia uses the package ```excatextractr``` which is available from CRAN. If you have not yet installed this package, please do so by typing the following command in your R console: 
-
-```{r, eval=FALSE}
-install.packages("exactextractr")
-```
-
-You can install the development version from [GitHub](https://github.com/){target="_blank"} with:
[…]
