+++ vignettes/01_description_raw_data.Rmd
@@ -1,0 +1,161 @@
+---
+title: "Description of data raw types"
+output: rmarkdown::html_vignette
+bibliography: [references.bib, HMA_bibliography.bib]
+vignette: >
+  %\VignetteIndexEntry{Description of raw data}
+  %\VignetteEngine{knitr::knitr}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+knitr::opts_knit$set(
+  rmarkdown.html_vignette.check_title = FALSE
+)
+```
+
+The package riversCentralAsia can read several types of raw data from time series in csv files to nc files, all of which are described in more detail below. 
+
+```{r, echo=FALSE, message=FALSE, error=FALSE, warning=FALSE}
+library(tidyverse)
+library(lubridate)
+library(timetk)
+library(riversCentralAsia)
+```
+
+## Time series
+In the former Soviet republics, river discharge but also climate variables like temperature and precipitation are often available from hydrological yearbooks published by the national HydroMeteorological institutes. The common data format for these time series has years in rows and months or decade (10-day interval) per year in columns as indicated in the two examples below. 
+
+```{r, echo=FALSE}
+knitr::kable(
+  tibble::tribble(
+  ~Year, ~Jan, ~Feb, ~Mar, ~..., ~Nov, ~Dec,   
+ "2001","0.1","0.1","...",  " ","...","0.2",  
+ "2001","0.1","0.2","...",  " ",  " ","...",
+ "2003","0.2","...",  " ",  " ",  " ",  " ",
+  "...","...",  " ",  " ",  " ",  " ",  " "), 
+ align = "c", format = "pipe", padding = 3, 
+ caption = "Example for monthly data format.")
+``` 
+or    
+   
+```{r, echo=FALSE}
+knitr::kable(
+  tibble::tribble(
+  ~Year, ~Jan, ~Jan, ~Jan, ~Feb, ~Feb, ~Feb, ~Mar, ~..., ~Dec, ~Dec,   
+    " ",  "I", "II","III",  "I", "II","III",  "I","...", "II","III",   
+ "2001","0.1","0.1","...",  " ",  " ",  " ",  " ",  " ","...","0.2",  
+ "2001","0.1","0.2","...",  " ",  " ",  " ",  " ",  " ",  " ","...",
+ "2003","0.2","...",  " ",  " ",  " ",  " ",  " ",  " ",  " ",  " ",
+  "...","...",  " ",  " ",  " ",  " ",  " ",  " ",  " ",  " ",  " "), 
+ align = "c", format = "pipe", padding = 2, 
+ caption = "Example for decadal data format.")
+```
+
+   
+To load the time series data using the function `loadTabularData`, save the cells without the headers in a csv file. We recommend commas as field separators and points as decimal separators.   
+
+A valid csv file example for monthly time series data is given below.   
+     
+      
+1990,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1,1.1,1.2    
+1991,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1,1.1,1.2    
+1992,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1,1.1,1.2    
+    
+    
+In the [article with the discharge processing example](https://hydrosolutions.github.io/riversCentralAsia/articles/01-discharge-processing-examples.html) we show how you can load a csv file and how to produce various visualizations using the riversCentralAsia package (and it's numerous dependencies).  
+
+## Climate raster data
[…]

[…]
