+++ vignettes/03-RSMinerveIO.md
@@ -1,247 +1,0 @@
----
-title: "I/O for RS Minerve"
-output: rmarkdown::html_vignette
-bibliography: [references.bib, HMA_bibliography.bib]
-vignette: >
-  %\VignetteIndexEntry{I/O for RS Minerve}
-  %\VignetteEngine{knitr::knitr}
-  %\VignetteEncoding{UTF-8}
----
-
-
-
-[RS Minerve](https://crealp.ch/rs-minerve/){target="_blank"} is a free to use hydrologic-hydraulic modelling software. It's accessibility makes is especially suitable for typical Hydromet tasks like flow forecasting. The package riversCentralAsia includes a number of functions to read and write files which interact with RS Minerve. 
-These tools are interesting for automated writing of forcing files and for reading post-processing and visualization of simulation results or model parameters. 
-
-Using the package `RSMinerveR`, the software can be run from inside R. 
-
-For details on model components please refer to the [RS Minerve manuals](https://crealp.github.io/rsminerve-releases/){target="_blank"}. 
-
-## Reading and writing parameters
-The following RSMinerve objects are currently supported in the `riversCentralAsia` package:\
-- Station\
-- Source\
-- SOCONT\
-- Kinematic\
-- Junction\
-- GSM\
-- HBV92\
-- Comparator
-
-The parameters for these objects are described in the [RS Minerve User Manual](https://crealp.github.io/rsminerve-releases/){target="_blank"}. 
-
-
-```r
-library(tidyverse)
-library(lubridate)
-library(riversCentralAsia)
-
-parameters_orig <- readRSMParameters("../../atbashy_glacier_demo_data/RSMINERVE/Atbaschy_PAR_postCal.txt") |> 
-  dplyr::filter(Parameters == "Ku [1/d]") 
-
-ggplot(parameters_orig) + 
-  geom_col(aes(Zone, Values), position = "dodge") + 
-  ylab(parameters_orig$Parameters[1]) + 
-  theme_bw()
-```
-
-![Storage coefficient of the upper reservoirs in the HBV models by zone.](figure/unnamed-chunk-2-1.png)
-
-The following shows an example how the model parameters can be changed and written to a RS Minerve-readable format. 
-
-
-```r
-# Doubble all Ku-values in zone A: 
-parameters <- readRSMParameters(
-  "../../atbashy_glacier_demo_data/RSMINERVE/Atbaschy_PAR_postCal.txt") |> 
-  mutate(Values = ifelse(Parameters == "Ku [1/d]" & Zone == "A", Values * 2, 
-                         Values))
-
-comparison <- parameters_orig |> 
-  mutate(Set = "Original") |> 
-  add_row(parameters |> dplyr::filter(Parameters == "Ku [1/d]") |> 
-            mutate(Set = "Modified"))
-
-ggplot(comparison) + 
-  geom_col(aes(Zone, Values, fill = Set), position = "dodge") + 
-  scale_fill_viridis_d() + 
-  ylab(parameters_orig$Parameters[1]) + 
-  theme_bw()
-```
-
-![plot of chunk unnamed-chunk-3](figure/unnamed-chunk-3-1.png)
-
-```r
-
-# Uncomment to write
-# writeRSMParameters(parameters, "Parameters_modified_Ku.txt")
-```
-
-## Reading simulation results
-
-Another useful application for the RS Minerve I/O functions is the reading of simulation results and the analysis of the partitioning of the total basin discharge into the individual fluxes. The code below illustrates how this may be done: 
-
-### Writing a check node file
[…]
