### 2022-10-14 ce7ef42
Delete 01-description_raw_data.knit.md
