+++ vignettes/01-description_raw_data.knit.md
@@ -1,149 +1,0 @@
----
-title: "Description of data raw types"
-output: rmarkdown::html_vignette
-bibliography: [references.bib, HMA_bibliography.bib]
-vignette: >
-  %\VignetteIndexEntry{Description of raw data}
-  %\VignetteEngine{knitr::knitr}
-  %\VignetteEncoding{UTF-8}
----
-
-
-
-The package riversCentralAsia can read several types of raw data from time series in csv files to nc files, all of which are described in more detail below. 
-
-
-
-## Time series
-In the former Soviet republics, river discharge but also climate variables like temperature and precipitation are often available from hydrological yearbooks published by the national HydroMeteorological institutes. The common data format for these time series has years in rows and months or decade (10-day interval) per year in columns as indicated in the two examples below. 
-
-
-Table: Example for monthly data format.
-
-|   Year   |   Jan   |   Feb   |   Mar   |   ...   |   Nov   |   Dec   |
-|:--------:|:-------:|:-------:|:-------:|:-------:|:-------:|:-------:|
-|   2001   |   0.1   |   0.1   |   ...   |         |   ...   |   0.2   |
-|   2001   |   0.1   |   0.2   |   ...   |         |         |   ...   |
-|   2003   |   0.2   |   ...   |         |         |         |         |
-|   ...    |   ...   |         |         |         |         |         |
-or    
-   
-
-Table: Example for decadal data format.
-
-|  Year  |  Jan  |  Jan  |  Jan  |  Feb  |  Feb  |  Feb  |  Mar  |  ...  |  Dec  |  Dec  |
-|:------:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|
-|        |   I   |  II   |  III  |   I   |  II   |  III  |   I   |  ...  |  II   |  III  |
-|  2001  |  0.1  |  0.1  |  ...  |       |       |       |       |       |  ...  |  0.2  |
-|  2001  |  0.1  |  0.2  |  ...  |       |       |       |       |       |       |  ...  |
-|  2003  |  0.2  |  ...  |       |       |       |       |       |       |       |       |
-|  ...   |  ...  |       |       |       |       |       |       |       |       |       |
-
-   
-To load the time series data using the function `loadTabularData`, save the cells without the headers in a csv file. We recommend commas as field separators and points as decimal separators.   
-
-A valid csv file example for monthly time series data is given below.   
-     
-      
-1990,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1,1.1,1.2    
-1991,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1,1.1,1.2    
-1992,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1,1.1,1.2    
-    
-    
-In the [article with the discharge processing example](https://hydrosolutions.github.io/riversCentralAsia/articles/01-discharge-processing-examples.html) we show how you can load a csv file and how to produce various visualizations using the riversCentralAsia package (and it's numerous dependencies).  
-
-## Climate raster data
[…]
