+++ DESCRIPTION
@@ -40,6 +40,7 @@ Imports:
     tibble,
     tidyr,
     tidyselect,
+    tidyverse,
     timetk,
     units,
     xml2,
