+++ pkgdown.yml
@@ -1,0 +1,13 @@
+pandoc: 2.14.2
+pkgdown: 2.0.6
+pkgdown_sha: ~
+articles:
+  00-introduction: 00-introduction.html
+  01-description_raw_data: 01-description_raw_data.html
+  01-discharge-processing-examples: 01-discharge-processing-examples.html
+  02-preparation-of-climate-forcing: 02-preparation-of-climate-forcing.html
+  03-RSMinerveIO: 03-RSMinerveIO.html
+  04-glacier-functions: 04-glacier-functions.html
+  05-snow-calibration: 05-snow-calibration.html
+last_built: 2022-10-14T16:26Z
+
