### 2022-10-17 #4805 @tonyewong
I fully support the no-laptop family holiday rules!
