### 2022-10-24 hydrosolutions/riversCentralAsia#109 @mengqi-z
Vignettes
Hi @mabesa,

This is for JOSS review https://github.com/openjournals/joss-reviews/issues/4805.

I went through the latest files within the vignettes folder, which is not yet successfully deployed and updated on the github documentation page. Anyway, vignettes covers a good amount of package functions. I enjoyed reading through all the examples. However, I suggest that all the data used in the example codes should be provided right before the codes, either directly as part of the R package data, or as downloadable links. Here are the issues I ran into:

1. In `vignettes/01-description_raw_dara.Rmd`, section "Climate Projections", the climate nc file provided for the example (https://www.dropbox.com/s/aw3mwh0ysydgu9j/tas_day_GFDL-ESM4_ssp126_r1i1p1f1_gr1.nc?dl=0) only covers data for 1979, where the code example shows 2012. Please either update the example code, or provide the correct file for the example.

2. In `vignettes/03-RSMinerveIO.Rmd`, section "Reading and writing parameters", the file `Atbaschy_PAR_postCal.txt` is not provided in the description. I later find this file in `atbashy_glacier_demo_data` folder mentioned in `04-glacier-functions.Rmd`, but it would be helpful to provide the dataset link every time any data is used in the example. For example, in section "Reading & plotting results" under `vignettes/03-RSMinerveIO.Rmd`, download link to `atbashy_glacier_demo_data` should be included.

3. In `vignettes/03-RSMinerveIO.Rmd`, section "Writing a check node file", variable `Object_IDs` is not defined in the example code.

4. In `vignettes/05-snow-calibration.Rmd`, section "Extract observed swe", `MA_SR_D_v01_N41_0E77_0_agg_16_WY1999_00_MASK.nc` and `MA_SR_D_v01_N41_0E77_0_agg_16_WY1999_00_SWE_SCA_POST.nc` are used in the example but not provided in `atbashy_glacier_demo_data/SNOW/`.

5. In `vignettes/05-snow-calibration.Rmd`, section "Extract observed swe", when running
`file2write <- file2write |> add_row(cbind(datesChar, data) |> mutate_all(as.character))`
I got this error:
`Error in data.frame(..., check.names = FALSE) : arguments imply differing number of rows: 5101, 40`.
I didn't get to run through the code section under "Extract observed swe" to produce `swel` due to reason 4. But I believe I am using `swel` from `SNOW/SWE.RData` loaded in the later section. So I am not sure why the length of `datesChar` and `data` are not the same. Please check and correct if needed.

Thank you.
