### 2022-10-24 #4805 @mengqi-z
Hi @mabesa - one more data point following the issue tonyewong mentioned previously, I also need to manually install package `exactextractr` before I can installed `riverCentralAsia` successfully. In addition, I posted an issue specifically for issues I encountered when going through vignettes https://github.com/hydrosolutions/riversCentralAsia/issues/109. Thanks!
