+++ pkgdown.yml
@@ -9,5 +9,5 @@ articles:
   03-RSMinerveIO: 03-RSMinerveIO.html
   04-glacier-functions: 04-glacier-functions.html
   05-snow-calibration: 05-snow-calibration.html
-last_built: 2022-10-24T07:30Z
+last_built: 2022-10-24T08:19Z
 
