### 2022-10-24 #4805 @mabesa
About the problem with the automatic installation of the exactextractr package: @mengqi-z and @tonyewong, could you please let me know: do you work on windows or another system? Was there a specific error message related to the exactextractr package? It always installs without problem on my windows machine. Maybe this is something I should take up with the developer.
