+++ README.Rmd
@@ -29,7 +29,7 @@ The most important functionalities of the package as well as the raw data that c
 
 Data preparation comes before hydrological modelling and is actually one of the biggest work chunks in the modelling process. This package includes a number of helper functions that can be connected to efficient workflows that automatize the data preparation process for hydrological modelling. The functionality includes: 
 
-- Efficient processing of present and future climate forcing, including hydro-meterological data from Central Asia ([time series](https://hydrosolutions.github.io/riversCentralAsia/articles/01-discharge-processing-examples.html){target="_blank"} and re-analysis data) and down-scaling of ERA5 re-analysis data (a more advanced topic which is described in the [course book](https://hydrosolutions.github.io/caham_book/climate_data.html){target="_blank"})
+- Efficient processing of present and future climate forcing, including hydro-meterological data from Central Asia ([time series and re-analysis data](https://github.com/hydrosolutions/riversCentralAsia/blob/master/.github/workflows/pkgdown.yaml){target="_blank"}) and down-scaling of ERA5 re-analysis data (a more advanced topic which is described in the [course book](https://hydrosolutions.github.io/caham_book/climate_data.html){target="_blank"})
 
 - The [preparation of GIS layers for automated model generation](https://hydrosolutions.github.io/riversCentralAsia/articles/02-preparation-of-climate-forcing.html){target="_blank"} and chapter [Geospatial data](https://hydrosolutions.github.io/caham_book/geospatial_data.html){target="_blank"} in the course book
 
@@ -66,7 +66,7 @@ If they are not yet available on your system, please install them using the foll
 ```{r, eval=FALSE}
 install.packages(c("rlang", "magrittr", "stringr", "purrr"))
 ```
-All other dependencies are installed automatically. 
+All other dependencies are installed automatically. Some users reported, that they had to install the package exactextractr manually. 
 
 Note that windows users require a working installation of [RTools](https://cran.r-project.org/bin/windows/Rtools/){target="_blank"} to install packages from github. 
 
