+++ vignettes/01-description_raw_data.Rmd
@@ -66,7 +66,7 @@ A valid csv file example for monthly time series data is given below.
 1991,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1,1.1,1.2    
 1992,0.1,0.2,0.3,0.4,0.5,0.6,0.7,0.8,0.9,1,1.1,1.2    
     
-    
+#### Example    
 In the [article with the discharge processing example](https://hydrosolutions.github.io/riversCentralAsia/articles/01-discharge-processing-examples.html) we show how you can load a csv file and how to produce various visualizations using the riversCentralAsia package (and it's numerous dependencies).  
 
 ## Climate raster data
@@ -79,13 +79,16 @@ NetCDF files can be quite large so it is often more convenient to use different 
 
 The CHELSA v2.1 NetCDF files are in the WGS84 projection with EPSG:4326. Please refer to the documentation of the data set you download. Ours includes daily precipitation sums in mm (varname = pr) or daily average near surface temperature in deg C (varname = tas) for each day of the year. There is one nc file for each year and for each climate variable. The date is included in the names of the daily images in the format X%Y.%m.%d (see example below).  
 
+#### Example
 To reproduce the examples below, please download the file [test_gen_HRU_Climate_CSV_RSMinerve_input_file.nc](https://github.com/hydrosolutions/riversCentralAsia/blob/master/tests/testthat/test_gen_HRU_Climate_CSV_RSMinerve_input_file.nc){target="_blank"} from the test folder in the riversCentralAsia package on github. 
 
 ```{r, eval=FALSE}
 library(raster)
 library(riversCentralAsia)
 
-# Adapt the path to the nc file
+# Download the example nc file from https://github.com/hydrosolutions/riversCentralAsia/blob/master/tests/testthat/test_gen_HRU_Climate_CSV_RSMinerve_input_file.nc
+
+# Adapt the path to the nc file on your computer
 ncfilepath = "tests/testthat/test_gen_HRU_Climate_CSV_RSMinerve_input_file.nc"
 
 # Read in the nc file as raster brick
@@ -121,10 +124,13 @@ The use of climate projection data for hydrological modelling is described in de
 
 Should you wish to download projections from further climate models, you can download the climate projection data for the CMIP 6 phase from the [Copernicus data repository app](https://cds.climate.copernicus.eu/cdsapp#!/dataset/projections-cmip6?tab=form){target="_blank"}. It is then recommended to inspect the downloaded data for any obvious problems using panoply and to reformat the data to a riversCentralAsia-package-readable format using cdo or R (or any other scripting language of your choosing). 
 
-To reproduce the example below, please download the [example nc file](https://www.dropbox.com/s/aw3mwh0ysydgu9j/tas_day_GFDL-ESM4_ssp126_r1i1p1f1_gr1.nc?dl=0){target="_blank"} from the data repository of the course book. It contains daily temperatures of the general circulation model (GCM) GFDL-ESM4 for the shared socio-economic scenario (SSP) 1 -2.6. Please refer to the [course book](https://hydrosolutions.github.io/caham_book/climate_data.html#sec-climate-projections){target="_blank"} for a more detailed explanation of GCMs and SSPs. 
+#### Example
[…]
