### 2022-10-24 6b1dd40
Fixed issues in vignette on RS MINERVE I/O
