### 2022-10-24 hydrosolutions/riversCentralAsia#109 @mabesa
Thank you very much for taking the time to go through the examples in such detail. Your input is very valuable!

The site is not successfully deployed again. I had a curl problem when compiling the site offline but through github actions everything is now working as it should.

(Not mentioned here but in https://github.com/openjournals/joss-reviews/issues/4805): I added a mention about the exactextractr package installation which has to be done manually in some systems. I cannot reproduce the need to manually install the package but I should probably follow up on this.

Issues with using external data for examples in the articles. I cannot include the examples as they are in the package because they are too big. The sources for the example data are not easily found in the example text so I added section headers for the examples.

1. The linked file you mention shows data from 2012 onwards (the example file in the previous section CHELSA covers data from 1979). I added the sources of the example files in the code chunks to make it a bit clearer how the examples can be reproduced.

2. I added the link to the file in the text directly above the examples and in the example chunks.

3. I added reading of the example file that was missing here. The example is now complete.

4. I see the two files in that folder. I will ask a colleague to check if he sees these files and will continue addressing the comments in a second comment.
