### 2022-10-24 391a3c2
Update 05-snow-calibration.rmd
