### 2022-10-24 #4805 @mabesa
A colleague of mine also kindly tested the clean-slate installation and didn't have issues with exactectractr. We also didn't encounter installation problems (other than missing RTools) with our students. I'm a bit stumped.
