### 2022-10-24 #4805 @mengqi-z
Hi @mabesa,

I am using MacOS Monterey. I think it might be the compilation issue, because I can install it separately when choosing to install from binary version (0.7.2), instead of the source version (0.9.0) which need compilation. In the latter case, `geos-config` does not exist and causes the problem. Here is the error message I got for the part where `exactextractr` failed to install during `riverCentralAsia` installation in the case of choosing "Yes" for the question "Do you want to install from sources the package which needs compilation? (Yes/no/cancel)":

```
* installing *source* package ‘exactextractr’ ...
** package ‘exactextractr’ successfully unpacked and MD5 sums checked
** using staged installation
configure: exactextractr: 0.9.0
checking for geos-config... no
configure: error: geos-config not found or not executable
ERROR: configuration failed for package ‘exactextractr’
* removing ‘/Library/Frameworks/R.framework/Versions/4.2-arm64/Resources/library/exactextractr’

The downloaded source packages are in
	‘/private/var/folders/0f/xg802qr91dd8gmmk8vwgsz340000gn/T/RtmpbGfVD1/downloaded_packages’
✔  checking for file ‘/private/var/folders/0f/xg802qr91dd8gmmk8vwgsz340000gn/T/RtmpbGfVD1/remotes173f7c639586/hydrosolutions-riversCentralAsia-5ee0d64/DESCRIPTION’ ...
─  preparing ‘riversCentralAsia’:
─  building ‘riversCentralAsia_1.1.0.tar.gz’

ERROR: dependency ‘exactextractr’ is not available for package ‘riversCentralAsia’
* removing ‘/Library/Frameworks/R.framework/Versions/4.2-arm64/Resources/library/riversCentralAsia’
Warning messages:
1: In i.p(...) :
  installation of package ‘exactextractr’ had non-zero exit status
2: In i.p(...) :
  installation of package ‘/var/folders/0f/xg802qr91dd8gmmk8vwgsz340000gn/T//RtmpbGfVD1/file173f66f10b30/riversCentralAsia_1.1.0.tar.gz’ had non-zero exit status
```
