### 2022-10-25 #4805 @mabesa
Thanks a lot! It is a Mac issue then (My colleague and me are working on Windows 11 at the moment). I have made a note in readme and written to the developer of exactextractr. I'll organize a mac for testing too and possibly should run a test on ubuntu as well...
