### 2022-11-04 #4805 @tonyewong
@mabesa Could a bit more be said in the Summary section of the paper about how this package might be useful in some broader applications, beyond specifically for students? I can appreciate that much of this was developed for a course, but I can also see how this sort of software will be useful for practitioners and researchers, in addition to students.

Then in Statement of Need/Related packages, it isn't immediately clear to me as a non-specialist how RS Minerve and the other mentioned packages fits into this picture in terms of functionality, as compared to the riversCentralAsia package. I was maybe a little confused here because the first part of Related Packages mentions the RSMinverveR package, which it says can run RS Minerve through R. But later this section says "there is no R package to facilitate hydrological modelling specifically with RS Minerve". There is probably a distinction here that I'm not quite getting. Can this part be clarified?

I'll open an issue in the repository for these, but wanted to mention it here first as well.
