### 2022-11-04 hydrosolutions/riversCentralAsia#112 @tonyewong
clarifications/additional details in JOSS paper
https://github.com/openjournals/joss-reviews/issues/4805#issuecomment-1302886413
