+++ README.md
@@ -1,259 +1,0 @@
-
-<!-- README.md is generated from README.Rmd. Please edit that file -->
-
-# riversCentralAsia
-
-## Summary
-
-The R package riversCentralAsia includes a set of tools to facilitate
-and automate data preparation for hydrological modelling. It thus
-contributes to more reproducible modeling workflows and makes
-hydrological modeling more accessible to students but also to interested 
-professional modellers. The figure below gives an overview over the 
-data perparation and modelling workflows supported by the package. 
-
-![Overview over the hydrological modelling workflow facitlitated by the riversCentralAsia package (Image source: Marti et al., 2022).](fig_01_model_chain_bw.png){width=60%}
-
-The package has been developed within the frame of a master level course
-on applied hydrological modelling in Central Asia and is extensively
-used in the open-source book
-<a href="https://hydrosolutions.github.io/caham_book/"
-target="_blank">Modeling of Hydrological Systems in Semi-Arid Central
-Asia</a> (Siegfried & Marti, 2022). The workflows are further validated within the
-Horizon 2020 project
-<a href="https://hydro4u.eu/" target="_blank">HYDRO4U</a>.
-
-While the package has been developed for the Central Asia region, most
-of the functions are generic and can be used for modelling projects
-anywhere in the world.
-
-The most important functionalities of the package as well as the raw
-data that can be processed with the package are described in the the
-articles of the
-<a href="https://hydrosolutions.github.io/riversCentralAsia/"
-target="_blank">project documentation site</a> but the examples in the
-book <a href="https://hydrosolutions.github.io/caham_book/"
-target="_blank">Modeling of Hydrological Systems in Semi-Arid Central
-Asia</a> demonstrate the full range of functions available and how to
-use them in a workflow.
-
-## Statement of need
-
-Data preparation comes before hydrological modelling and is actually one
-of the biggest work chunks in the modelling process. This package
-includes a number of helper functions that can be connected to efficient
-workflows that automatize the data preparation process for hydrological
-modelling. The functionality includes:
-
--   Efficient processing of present and future climate forcing,
-    including hydro-meterological data from Central Asia (<a
-    href="https://hydrosolutions.github.io/riversCentralAsia/articles/01-discharge-processing-examples.html"
-    target="_blank">time series</a> and re-analysis data) and
-    down-scaling of ERA5 re-analysis data (a more advanced topic which
-    is described in the
-    <a href="https://hydrosolutions.github.io/caham_book/climate_data.html"
-    target="_blank">course book</a>)
-
--   The <a
-    href="https://hydrosolutions.github.io/riversCentralAsia/articles/02-preparation-of-climate-forcing.html"
-    target="_blank">preparation of GIS layers for automated model
-    generation</a> and chapter <a
-    href="https://hydrosolutions.github.io/caham_book/geospatial_data.html"
-    target="_blank">Geospatial data</a> in the course book
-
--   <a
-    href="https://hydrosolutions.github.io/riversCentralAsia/articles/04-glacier-functions.html"
-    target="_blank">Volume area scaling of glaciers</a>
-
--   Post-processing of simulation results, e.g. <a
-    href="https://hydrosolutions.github.io/riversCentralAsia/articles/05-snow-calibration.html"
-    target="_blank">extraction and visualisation of snow water
-    equivalent</a> or <a
-    href="https://hydrosolutions.github.io/riversCentralAsia/reference/computeAnnualFlowDurationCurve.html"
-    target="_blank">computation of flow duration curves</a>
-
--   I/O interface with the hydrologic-hydraulic modelling software
[…]
