### 2022-11-04 06ca093
Reverted changes on md and made them on rmd
