### 2022-11-04 20e14d6
update readme (add illustration of workflow)
