### 2022-11-04 hydrosolutions/riversCentralAsia#112 @mabesa
Thank you for your persistence! I did a minimal update of both the readme in the main branch and the paper (based on the readme) in the paper branch and re-compiled the paper.
Could you let me know if the points are clarified or if I should invest more words into describing the context of the functions in the paper?
