+++ paper.md
@@ -1,0 +1,189 @@
+---
+title: 'riversCentralAsia: An R package for hydrological modelling'
+tags:
+- R
+- hydrology
+- modelling
+- Central Asia
+date: "3 August 2022"
+output: pdf_document
+authors:
+- name: Beatrice Marti
+  orcid: "0000-0003-2089-3478"
+  equal-contrib: yes
+  affiliation: 1
+- name: Tobias Siegfried
+  orcid: "0000-0002-2995-9253"
+  equal-contrib: yes
+  affiliation: 1
+bibliography: paper.bib
+affiliations:
+- name: hydrosolutions GmbH, Zurich, Switzerland
+  index: 1
+---
+
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+
+
+## Summary
+
+The R package riversCentralAsia includes a set of tools to facilitate and automate data preparation for hydrological modelling. It thus contributes to more reproducible modeling workflows and makes hydrological modeling more accessible to students and to interested professional modelers. 
+
+The package has been developed within the frame of a master level course on applied hydrological modelling in Central Asia and is extensively used in the open-source book Modeling of Hydrological Systems in Semi-Arid Central Asia [@CAHAM:2022]. The workflows are further validated within the Horizon 2020 project HYDRO4U [@Hydro4u:2022].
+
+While the package has been developed for the Central Asia region, most of the functions are generic and can be used for modelling projects anywhere in the world.
+
+The most important functionalities of the package as well as the raw data that can be processed with the package are described in the the articles of the [project documentation site](https://hydrosolutions.github.io/riversCentralAsia/){target="_blank"} but the examples in the course book [Modeling of Hydrological Systems in Semi-Arid Central Asia](https://hydrosolutions.github.io/caham_book/){target="_blank"} demonstrate the full range of functions available and how touse them in a workflow.
+
+## Statement of need
+
+Data preparation comes before hydrological modelling and is actually one
+of the biggest work chunks in the modelling process. This package
+includes a number of helper functions that can be connected to efficient
+workflows that automatize the data preparation process for hydrological
+modelling as shown in the figure below. 
+
+```{r out.width='80%', echo=FALSE}
+knitr::include_graphics("./man/figures/fig_01_model_chain_bw.png")
+```
+  
+Figure: Overview over the modelling workflow supported by the riversCentralAsia package (Image source: Marti et al., 2022). Abbreviations are explained below. The workflow relies entirely on free, publicly available data and software.    
+
+The data preparation step includes the delineation of river catchment boundaries and the derivation of hydrological response units (HRU) using gauge locations, the [SRTM digital elevation model (DEM)](https://doi.org/10.5067/MEaSUREs/SRTM/SRTMGL1.003){target="_blank"} and glacier outlines from the [Randolph Glacier Inventory (RGI)](https://doi.org/10.7265/N5-RGI-60){target="_blank"}.
+
[…]
