### 2022-11-04 #4805 @mabesa
I can't get the .github/workflows/draft-pdf.yml workflow to compile the current version of the paper branch. I can only re-run the last compilation which was sometime in October. paper.md is updated but I'm not sure the paper pdf is beeing updated for joss as it should be.
