### 2022-11-04 3ce7ebf
Update figure command
