### 2022-11-04 #4805 @mabesa
@tonyewong thanks for the hint with the editorialbot command. The figure now shows.
So: I did a (admittedly very minimalistic) update of the paper text by adding a very rough description of the hydrological modelling workflow where the package is used. Please let me know if this makes the package purpose more understandable or if it adds to the confusion. I could free up some time next week to restructure the paper text.
