### 2022-11-04 #4805 @tonyewong
Looks good! The one point that I think might still be a bit unclear is the relationship to the RSMinverveR package:

> Then in Statement of Need/Related packages, it isn't immediately clear to me as a non-specialist how RS Minerve and the other mentioned packages fits into this picture in terms of functionality, as compared to the riversCentralAsia package. I was maybe a little confused here because the first part of Related Packages mentions the RSMinverveR package, which it says can run RS Minerve through R. But later this section says "there is no R package to facilitate hydrological modelling specifically with RS Minerve". There is probably a distinction here that I'm not quite getting. Can this part be clarified?
