### 2022-11-07 b781a24
Adapted the "related packages" section in paper
Clarification of difference between riversCentralAsia and RSMinerveR.
