### 2022-11-08 #4805 @mabesa
I'm finally starting to see your point. RSMinerveR is not a real software package. It includes some of the functions most of which I copied to riversCentralAsia later so they are in fact redundant. The essential part in RSMinerveR are the examples/vignettes that demonstrate how to use the CLR interface to run RS MINERVE from R without using the GUI of RS MINERVE. I'll rephrase the sections to clarify that and I'll clean out redundancies from RSMinerveR as soon as I have time!
