### 2022-11-08 #4805 @mabesa
both riversCentralAsia and RSMinerveR are update. RSMinerveR is still organised as a package but it now only contains the examples as vignettes, no redundant R functions. The examples use the functions from riversCentralAsia.
