### 2022-11-15 dca763e
Update r_version_tests.yml
Add testing version 4.2.2
