### 2022-11-15 7ca706e
Update r_version_tests.yml
Problem with exactextractr, not available for 4.2.1
