+++ .github/workflows/r_version_tests.yml
@@ -22,7 +22,7 @@ jobs:
     runs-on: macos-latest
     strategy:
       matrix:
-        r-version: ['4.2.1', '4.2.2']
+        r-version: ['4.2.2']
 
     steps:
       - uses: actions/checkout@v3
