### 2022-11-16 hydrosolutions/riversCentralAsia#116 @mengqi-z
Software Paper Review
Hi @mabesa,

I have few feedback on the software paper for JOSS https://github.com/openjournals/joss-reviews/issues/4805

1. Title - the title misleads the reader to believe this package performs some sort of hydrological modeling. However, the package mainly facilitates data preparation and processing. Please refine the title to reflect the package's actual functionality.

2. Line 8 - please be consistent and use either "modelling" or "modeling", not both.

3. Statement of need
- The second paragraph in "statement of need" states the data preparation steps. However, it includes steps that are not actually done in the `riversCentralAsia` package, which could mislead readers on the package functionality. For example, I believe "delineation of river catchment boundaries" is processed in QGIS. Please revise the paragraph and focus on what functions this package can deliver.
- The third paragraph in "statement of need" is great material for data availability, but it does not extend to how the package is needed to process those data. I would suggest adding one or two sentences to highlight the ease-to-use feature of the package to process those datasets.
- Line 41 mentioned down-scaling of ERA5 re-analysis data. I think this is pointing to `doQuantileMapping` function. I am not quite sure if this function fits the concept of  "downscaling" because the function is more like bias-correction. Also, downscaling can include spatial and temporal downscaling.

4. Installation - could you include the R version (>= 4.1) requirement for this package? Or at least add this requirement to the Github Page.

5. Figure 1 - some abbreviations are not explained in the figure or paper (e.g., GSM).
