### 2022-11-16 hydrosolutions/riversCentralAsia#117 @mengqi-z
[Community Guidelines] Add Contributor Guidelines
Hi @mabesa,

In Community Guidelines section, please also provide "Contributor Guidelines" for those who wish to contribute to the package. This can be added as a CONTRIBUTING.md file or directly under the Community Guidelines section.  If you wish to add CONTRIBUTING.md, here is an [example](https://github.com/rails/rails/blob/main/CONTRIBUTING.md).

For JOSS review https://github.com/openjournals/joss-reviews/issues/4805
