### 2022-11-16 #4805 @mengqi-z
Hi @mabesa,

I opened several issues regarding the software paper, community guidelines, and functionality. I still have few functions from the package to checkout. But I am getting close to be done with the review. Great work!
