+++ CONTRIBUTING.md
@@ -1,0 +1,33 @@
+## How to contribute to riversCentralAsia
+
+#### **Did you find a bug?**
+
+* **Ensure the bug was not already reported** by searching on GitHub under [Issues](https://github.com/hydrosolutions/riversCentralAsia/issues).
+
+* If you're unable to find an open issue addressing the problem, [open a new one](https://github.com/hydrosolutions/riversCentralAsia/issues/new). Be sure to include a **title and clear description**, as much relevant information as possible, and a **code sample** or an **executable test case** demonstrating the expected behavior that is not occurring.
+
+#### **Did you write a patch that fixes a bug?**
+
+* Open a new GitHub pull request with the patch.
+
+* Ensure the PR description clearly describes the problem and solution. Include the relevant issue number if applicable.
+
+#### **Did you fix whitespace, format code, or make a purely cosmetic patch?**
+
+Changes that are cosmetic in nature and do not add anything substantial to the stability, functionality, or testability of riversCentralAsia will generally not be accepted.
+
+#### **Do you intend to add a new feature or change an existing one?**
+
+* Contact the developers at marti at hydrosolutions point ch to discuss your intentions. 
+
+#### **Do you have questions about the source code?**
+
+* Ask any question about how to use riversCentralAsia to marti at hydrosolutions point ch. 
+
+#### **Do you want to contribute to the Rails documentation?**
+
+* Contributions to the documentation are very welcome. Please open an issue with suggestions for improvements or open a pull request with your suggested documentation changes.  
+
+Thanks! 
+
+The riversCentralAsia-Team at hydrosolutions
