+++ README.md
@@ -85,38 +85,38 @@ Some alternative geoprocessing workflows are described in
 
 The riversCentralAsia package functionality includes:
 
--   Efficient processing of present and future climate forcing,
-    including hydro-meterological data from Central Asia (<a
-    href="https://hydrosolutions.github.io/riversCentralAsia/articles/01-discharge-processing-examples.html"
-    target="_blank">time series</a> and re-analysis data) and
-    down-scaling of ERA5 re-analysis data (a more advanced topic which
-    is described in the
-    <a href="https://hydrosolutions.github.io/caham_book/climate_data.html"
-    target="_blank">course book</a>)
+- Efficient processing of present and future climate forcing, including
+  hydro-meterological data from Central Asia (<a
+  href="https://hydrosolutions.github.io/riversCentralAsia/articles/01-discharge-processing-examples.html"
+  target="_blank">time series</a> and re-analysis data) and down-scaling
+  of ERA5 re-analysis data (a more advanced topic which is described in
+  the
+  <a href="https://hydrosolutions.github.io/caham_book/climate_data.html"
+  target="_blank">course book</a>)
 
--   The <a
-    href="https://hydrosolutions.github.io/riversCentralAsia/articles/02-preparation-of-climate-forcing.html"
-    target="_blank">preparation of GIS layers for automated model
-    generation</a> and chapter <a
-    href="https://hydrosolutions.github.io/caham_book/geospatial_data.html"
-    target="_blank">Geospatial data</a> in the course book
+- The <a
+  href="https://hydrosolutions.github.io/riversCentralAsia/articles/02-preparation-of-climate-forcing.html"
+  target="_blank">preparation of GIS layers for automated model
+  generation</a> and chapter <a
+  href="https://hydrosolutions.github.io/caham_book/geospatial_data.html"
+  target="_blank">Geospatial data</a> in the course book
 
--   <a
-    href="https://hydrosolutions.github.io/riversCentralAsia/articles/04-glacier-functions.html"
-    target="_blank">Volume area scaling of glaciers</a>
+- <a
+  href="https://hydrosolutions.github.io/riversCentralAsia/articles/04-glacier-functions.html"
+  target="_blank">Volume area scaling of glaciers</a>
 
--   Post-processing of simulation results, e.g. <a
-    href="https://hydrosolutions.github.io/riversCentralAsia/articles/05-snow-calibration.html"
-    target="_blank">extraction and visualisation of snow water
-    equivalent</a> or <a
-    href="https://hydrosolutions.github.io/riversCentralAsia/reference/computeAnnualFlowDurationCurve.html"
-    target="_blank">computation of flow duration curves</a>
+- Post-processing of simulation results, e.g. <a
+  href="https://hydrosolutions.github.io/riversCentralAsia/articles/05-snow-calibration.html"
+  target="_blank">extraction and visualisation of snow water
+  equivalent</a> or <a
+  href="https://hydrosolutions.github.io/riversCentralAsia/reference/computeAnnualFlowDurationCurve.html"
+  target="_blank">computation of flow duration curves</a>
 
--   I/O interface with the hydrologic-hydraulic modelling software that
-    allows reading and writing of input and output files of the
-    hydraulic-hydrological modelling software
-    <a href="https://crealp.ch/rs-minerve/" target="_blank">RS
-    Minerve</a>.
+- I/O interface with the hydrologic-hydraulic modelling software that
+  allows reading and writing of input and output files of the
+  hydraulic-hydrological modelling software
+  <a href="https://crealp.ch/rs-minerve/" target="_blank">RS
+  Minerve</a>.
 
 While here, we focus on the description of the individual functions, the
[…]
