+++ CONTRIBUTING.md
@@ -1,3 +1,4 @@
+# Contributing guidelines
 ## How to contribute to riversCentralAsia
 
 #### **Did you find a bug?**
