### 2022-11-17 hydrosolutions/riversCentralAsia#117 @mabesa
Done, thank you for the suggestion and the example!
