+++ paper.bib
@@ -1,3 +1,37 @@
+@Manual{qmap_package,
+    title = {qmap: Statistical transformations for post-processing
+      climate model output},
+    author = {Lukas Gudmundsson},
+    year = {2016},
+    note = {R package version 1.0-4},
+  }
+
+@article{gudmundsson_technical_2012,
+  title = {Technical {{Note}}: {{Downscaling RCM}} Precipitation to the Station Scale Using Statistical Transformations \&ndash; a Comparison of Methods},
+  shorttitle = {Technical {{Note}}},
+  author = {Gudmundsson, L. and Bremnes, J. B. and Haugen, J. E. and {Engen-Skaugen}, T.},
+  year = {2012},
+  month = sep,
+  journal = {Hydrology and Earth System Sciences},
+  volume = {16},
+  number = {9},
+  pages = {3383--3390},
+  publisher = {{Copernicus GmbH}},
+  issn = {1027-5606},
+  doi = {10.5194/hess-16-3383-2012},
+  langid = {english}
+}
+
+@article{egusphere-2022-966,
+  title = {A Comprehensive Open-Source Course for Teaching Applied Hydrological Modelling in {{Central Asia}}},
+  author = {Marti, B. S. and Zhumabaev, A. and Siegfried, T.},
+  year = {2022},
+  journal = {EGUsphere},
+  volume = {2022},
+  pages = {1--18},
+  doi = {10.5194/egusphere-2022-966}
+}
+
 @book{CAHAM:2022,
   title = {{Modeling of Hydrological Systems in Semi-Arid Central Asia (\url{https://hydrosolutions.github.io/caham_book/})}},
   author = {Siegfried, Tobias and Marti, Beatrice},
