+++ paper.md
@@ -26,7 +26,7 @@ affiliations:
 
 ## Summary
 
-The R package riversCentralAsia includes a set of tools to facilitate and automate data preparation for hydrological modelling. It thus contributes to more reproducible modelling workflows and makes hydrological modelling more accessible to students and to interested professional modelers. 
+The R package *riversCentralAsia* includes a set of tools to facilitate and automate data preparation for hydrological modelling. It thus contributes to more reproducible modelling workflows and makes hydrological modelling more accessible to students and to interested professional modelers. 
 
 The package has been developed within the frame of a master level course on applied hydrological modelling in Central Asia and is extensively used in the open-source book modelling of Hydrological Systems in Semi-Arid Central Asia [@CAHAM:2022]. The workflows are further validated within the Horizon 2020 project HYDRO4U [@Hydro4u:2022].
 
@@ -40,11 +40,11 @@ Data preparation comes before hydrological modelling and is actually one of the 
 
 !["Overview over the modelling workflow supported by the riversCentralAsia package (Image source: Marti et al., 2022). This package supports the data processing components in R. Abbreviations are explained in the text. The workflow relies entirely on free, publicly available data and software."](./man/figures/fig_01_model_chain_bw.png){ width=80% }  
 
-The data preparation step covered by the riversCentralAsia package the derivation of hydrological response units (HRU) using a basin outline and the [SRTM digital elevation model (DEM)](https://doi.org/10.5067/MEaSUREs/SRTM/SRTMGL1.003){target="_blank"}. The derivation of the basin outline and processing of geospatial layers for import to RS MINERVE in QGIS is described in detail in @CAHAM:2022.  
+The data preparation step covered by the *riversCentralAsia* package the derivation of hydrological response units (HRU) using a basin outline and the [SRTM digital elevation model (DEM)](https://doi.org/10.5067/MEaSUREs/SRTM/SRTMGL1.003){target="_blank"}. The derivation of the basin outline and processing of geospatial layers for import to RS MINERVE in QGIS is described in detail in @CAHAM:2022.  
 
 Although the High Mountain region of Central Asia is generally perceived as a data scarce region, a number of gridded data products are available that form a fair basis for regional hydrological modelling at seasonal time scales. [CHELSA v2.1](https://chelsa-climate.org/){target="_blank"} is a weather data product at 1 km2 by 1km2 resolution. The *riversCentralAsia* package includes the function [*gen_HRU_Climate_CSV_RSMinerve*](https://hydrosolutions.github.io/riversCentralAsia/reference/gen_HRU_Climate_CSV_RSMinerve.html){target="_blank"} which extracts CHELSA precipitation or temperature on the hydrological response units and returns the data in an RS MINERVE readable format.  
[…]
