+++ paper.md
@@ -26,9 +26,9 @@ affiliations:
 
 ## Summary
 
-The R package *riversCentralAsia* includes a set of tools to facilitate and automate data preparation for hydrological modelling. It thus contributes to more reproducible modelling workflows and makes hydrological modelling more accessible to students and to interested professional modelers. 
+The R package *riversCentralAsia* includes a set of tools to facilitate and automate data preparation for hydrological modelling. It thus contributes to more reproducible modelling workflows and makes hydrological modelling more accessible to students and to interested professional modellers. 
 
-The package has been developed within the frame of a master level course on applied hydrological modelling in Central Asia and is extensively used in the open-source book modelling of Hydrological Systems in Semi-Arid Central Asia [@CAHAM:2022]. The workflows are further validated within the Horizon 2020 project HYDRO4U [@Hydro4u:2022].
+The package has been developed within the frame of a master level course on applied hydrological modelling in Central Asia [@egusphere-2022-966] and is extensively used in the open-source book "Modelling of Hydrological Systems in Semi-Arid Central Asia" [@CAHAM:2022]. The workflows are further validated within the Horizon 2020 project HYDRO4U [@Hydro4u:2022].
 
 While the package has been developed for the Central Asia region, most of the functions are generic and can be used for modelling projects anywhere in the world.
 
@@ -36,58 +36,44 @@ The most important functionalities of the package as well as the raw data that c
 
 ## Statement of need
 
-Data preparation comes before hydrological modelling and is actually one of the biggest work chunks in the modelling process. This package includes a number of helper functions that can be connected to efficient workflows that automatize the data preparation process for hydrological modelling as shown in the figure below. 
+Data preparation comes before hydrological modelling and is actually one of the biggest work chunks in the modelling process. This package includes a number of helper functions that can be connected to efficient workflows that automatize the data preparation process for hydrological modelling as shown in the figure below. This supports a more reproducible modelling workflow and improves the scalability of hydrological modelling. 
 
-!["Overview over the modelling workflow supported by the riversCentralAsia package (Image source: Marti et al., 2022). This package supports the data processing components in R. Abbreviations are explained in the text. The workflow relies entirely on free, publicly available data and software."](./man/figures/fig_01_model_chain_bw.png){ width=80% }  
+!["Overview over the modelling workflow whos R-components are supported by the riversCentralAsia package (Image source: Marti et al., 2022). Abbreviations are explained in the text. The workflow relies entirely on free, publicly available data and software."](./man/figures/fig_01_model_chain_bw.png){ width=80% }  
 
-The data preparation step covered by the *riversCentralAsia* package the derivation of hydrological response units (HRU) using a basin outline and the [SRTM digital elevation model (DEM)](https://doi.org/10.5067/MEaSUREs/SRTM/SRTMGL1.003){target="_blank"}. The derivation of the basin outline and processing of geospatial layers for import to RS MINERVE in QGIS is described in detail in @CAHAM:2022.  
[…]
