### 2022-11-17 hydrosolutions/riversCentralAsia#116 @mabesa
Thanks a lot for your critical review. This improves the paper so much!
1. Title is adapted and should reflect the data processing focus of the package now
2. consistent use of modelling
3. Adapting statement of need
- I tried to more clearly state which components of the modelling workflow are covered by the package.
- A good point. I added examples of how the package supports the use of this data.
- quantile mapping is indeed a bias correction algorithm but it is used for statistical downscaling. I added a sentence on this and an appropriate reference.
4. Added R >= 4.1 under installation
5. Added a bracket and links for GSM and HBV
