### 2022-11-17 #4805 @mabesa
Dear @mengqi-z, thank you very much for your critical review! I addressed the issues and am now updating the README in the main repository to include the suggestions you've made for the paper.

@crvernon, @tonyewong and @mengqi-z I acknowledged your input in the README but I would like to acknowledge your input in the paper as well. Should I use your github user names?

Sorry to all for the multiple postings to editorialbot. I'm guessing you receive a couple of email notifications every time I recompile...
