+++ README.Rmd
@@ -44,24 +44,23 @@ use them in a workflow.
 
 ## Statement of need
 
-Data preparation comes before hydrological modelling and is actually one
-of the biggest work chunks in the modelling process. This package
-includes a number of helper functions that can be connected to efficient
-workflows that automatize the data preparation process for hydrological
-modelling as shown in the figure below. 
+Data preparation comes before hydrological modelling and is actually one of the biggest work chunks in the modelling process. This package includes a number of helper functions that can be connected to efficient workflows that automatize the data preparation process for hydrological modelling as shown in the figure below. This supports a more reproducible modelling workflow and improves the scalability of hydrological modelling.  
 
 ```{r out.width='80%', echo=FALSE}
 knitr::include_graphics("./man/figures/fig_01_model_chain_bw.png")
 ```
   
-Figure: Overview over the modelling workflow supported by the riversCentralAsia package (Image source: Marti et al., 2022). Abbreviations are explained below. The workflow relies entirely on free, publicly available data and software.    
+Figure: Overview over the modelling workflow whos R-components are supported by the riversCentralAsia package (Image source: Marti et al., 2022). Abbreviations are explained in the text. The workflow relies entirely on free, publicly available data and software.    
 
-The data preparation step includes the delineation of river catchment boundaries and the derivation of hydrological response units (HRU) using gauge locations, the [SRTM digital elevation model (DEM)](https://doi.org/10.5067/MEaSUREs/SRTM/SRTMGL1.003){target="_blank"} and glacier outlines from the [Randolph Glacier Inventory (RGI)](https://doi.org/10.7265/N5-RGI-60){target="_blank"}.
+The data preparation step covered by the *riversCentralAsia* package the derivation of hydrological response units (HRU) using a basin outline and the [SRTM digital elevation model (DEM)](https://doi.org/10.5067/MEaSUREs/SRTM/SRTMGL1.003){target="_blank"}. The [derivation of the basin outline](https://hydrosolutions.github.io/caham_book/geospatial_data.html#sec-catchment-delineation){target="_blank"} and [processing of geospatial layers for import to RS MINERVE](https://hydrosolutions.github.io/caham_book/geospatial_data.html#sec-preparation-of-rsminerve-input-files){target="_blank"} in QGIS is described in detail in @CAHAM:2022.  
 
-Although the High Mountain region of Central Asia is generally perceived as a data scarce region, a number of gridded data products are available that form a fair basis for regional hydrological modelling at seasonal time scales. [CHELSA v2.1](https://chelsa-climate.org/){target="_blank"} is a weather data product at 1 km2 by 1km2 resolution. [Glacier thinning](https://www.nature.com/articles/s41586-021-03436-z){target="_blank"} and [glacier ablation](https://doi.org/10.1038/s41467-021-23073-4){target="_blank"} are data sets from open-access literature. Data on snow water equivalents is sourced from the [High Mountain Asia Snow Reanalysis (HMASR) Product](https://doi.org/10.5067/HNAUGJQXSCVU){target="_blank"}, river discharge is taken from the hydrological year books of the HydroMeteorological Institutes in Central Asia, and [CMIP6 climate model results area available from Copernicus](https://cds.climate.copernicus.eu/cdsapp#!/dataset/projections-cmip6?tab=form){target="_blank"}. 
[…]
