+++ README.md
@@ -40,46 +40,79 @@ Data preparation comes before hydrological modelling and is actually one
 of the biggest work chunks in the modelling process. This package
 includes a number of helper functions that can be connected to efficient
 workflows that automatize the data preparation process for hydrological
-modelling as shown in the figure below.
+modelling as shown in the figure below. This supports a more
+reproducible modelling workflow and improves the scalability of
+hydrological modelling.
 
 <img src="./man/figures/fig_01_model_chain_bw.png" width="80%" />
 
-Figure: Overview over the modelling workflow supported by the
-riversCentralAsia package (Image source: Marti et al., 2022).
-Abbreviations are explained below. The workflow relies entirely on free,
-publicly available data and software.
+Figure: Overview over the modelling workflow whos R-components are
+supported by the riversCentralAsia package (Image source: Marti et al.,
+2022). Abbreviations are explained in the text. The workflow relies
+entirely on free, publicly available data and software.
 
-The data preparation step includes the delineation of river catchment
-boundaries and the derivation of hydrological response units (HRU) using
-gauge locations, the
-<a href="https://doi.org/10.5067/MEaSUREs/SRTM/SRTMGL1.003"
-target="_blank">SRTM digital elevation model (DEM)</a> and glacier
-outlines from the
-<a href="https://doi.org/10.7265/N5-RGI-60" target="_blank">Randolph
-Glacier Inventory (RGI)</a>.
+The data preparation step covered by the *riversCentralAsia* package the
+derivation of hydrological response units (HRU) using a basin outline
+and the <a href="https://doi.org/10.5067/MEaSUREs/SRTM/SRTMGL1.003"
+target="_blank">SRTM digital elevation model (DEM)</a>. The <a
+href="https://hydrosolutions.github.io/caham_book/geospatial_data.html#sec-catchment-delineation"
+target="_blank">derivation of the basin outline</a> and <a
+href="https://hydrosolutions.github.io/caham_book/geospatial_data.html#sec-preparation-of-rsminerve-input-files"
+target="_blank">processing of geospatial layers for import to RS
+MINERVE</a> in QGIS is described in detail in @CAHAM:2022.
 
 Although the High Mountain region of Central Asia is generally perceived
 as a data scarce region, a number of gridded data products are available
 that form a fair basis for regional hydrological modelling at seasonal
 time scales.
 <a href="https://chelsa-climate.org/" target="_blank">CHELSA v2.1</a> is
-a weather data product at 1 km2 by 1km2 resolution.
+a weather data product at 1 km2 grid resolution. The *riversCentralAsia*
+package includes the function <a
+href="https://hydrosolutions.github.io/riversCentralAsia/reference/gen_HRU_Climate_CSV_RSMinerve.html"
+target="_blank"><em>gen_HRU_Climate_CSV_RSMinerve</em></a> which
+extracts CHELSA precipitation or temperature on the hydrological
+response units and returns the data in an RS MINERVE readable format.  
 <a href="https://www.nature.com/articles/s41586-021-03436-z"
 target="_blank">Glacier thinning</a> and
 <a href="https://doi.org/10.1038/s41467-021-23073-4"
 target="_blank">glacier ablation</a> are data sets from open-access
-literature. Data on snow water equivalents is sourced from the
+literature which can be used to calibrate the <a
+href="https://crealp.github.io/rsminerve-releases/tech_hydrological_models.html#sec-model_gsm"
+target="_blank">GSM model objects</a> (a glacier runoff model) in RS
+MINERVE. Data on snow water equivalents is sourced from the
 <a href="https://doi.org/10.5067/HNAUGJQXSCVU" target="_blank">High
[…]
