### 2022-11-17 hydrosolutions/riversCentralAsia#116 @mengqi-z
Hi @mabesa,

Thank you for addressing my comments. I read through the latest version of the paper and I think it improved a lot. Some minor typos I noticed in the paper:

1. Line 2: there should be a space between "pre-" and "and".
2. Line 21: "touse" should be "to use".
3. Line 28: add "includes" after "package"?

I may not capture every typo. I suggest carefully read through the paper to make sure there are no typos and everything reads good. Thank you.
