+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'riversCentralAsia: An R package to support data pre-and postprocessing for hydrological modelling with RS MINERVE'
+title: 'riversCentralAsia: An R package to support data pre- and postprocessing for hydrological modelling with RS MINERVE'
 tags:
 - R
 - hydrology
@@ -32,27 +32,29 @@ The package has been developed within the frame of a master level course on appl
 
 While the package has been developed for the Central Asia region, most of the functions are generic and can be used for modelling projects anywhere in the world.
 
-The most important functionalities of the package as well as the raw data that can be processed with the package are described in the the articles of the [project documentation site](https://hydrosolutions.github.io/riversCentralAsia/){target="_blank"} but the examples in the course book [modelling of Hydrological Systems in Semi-Arid Central Asia](https://hydrosolutions.github.io/caham_book/){target="_blank"} demonstrate the full range of functions available and how touse them in a workflow.
+The most important functionalities of the package as well as the data that can be processed with the package are described in the the articles of the [project documentation site](https://hydrosolutions.github.io/riversCentralAsia/){target="_blank"} but the examples in the course book ["Modelling of Hydrological Systems in Semi-Arid Central Asia"](https://hydrosolutions.github.io/caham_book/){target="_blank"} demonstrate the full range of functions available and how to use them in a workflow.
 
 ## Statement of need
 
-Data preparation comes before hydrological modelling and is actually one of the biggest work chunks in the modelling process. This package includes a number of helper functions that can be connected to efficient workflows that automatize the data preparation process for hydrological modelling as shown in the figure below. This supports a more reproducible modelling workflow and improves the scalability of hydrological modelling. 
+Data preparation comes before hydrological modelling and is actually one of the biggest work chunks in the modelling process. This package includes a number of helper functions that can be connected to efficient workflows that automatize the data preparation process for hydrological modelling as shown in the figure below. The package thereby supports a more reproducible modelling workflow and improves the scalability of hydrological modelling. 
 
 !["Overview over the modelling workflow whos R-components are supported by the riversCentralAsia package (Image source: Marti et al., 2022). Abbreviations are explained in the text. The workflow relies entirely on free, publicly available data and software."](./man/figures/fig_01_model_chain_bw.png){ width=80% }  
 
-The data preparation step covered by the *riversCentralAsia* package the derivation of hydrological response units (HRU) using a basin outline and the [SRTM digital elevation model (DEM)](https://doi.org/10.5067/MEaSUREs/SRTM/SRTMGL1.003){target="_blank"}. The [derivation of the basin outline](https://hydrosolutions.github.io/caham_book/geospatial_data.html#sec-catchment-delineation){target="_blank"} and [processing of geospatial layers for import to RS MINERVE](https://hydrosolutions.github.io/caham_book/geospatial_data.html#sec-preparation-of-rsminerve-input-files){target="_blank"} in QGIS is described in detail in @CAHAM:2022.  
[…]
