### 2022-12-01 #4805 @tonyewong
Thanks for the reminder! Just getting things back on the rails after one holiday and before another...

I'm just finishing now going through some of the examples in the documentation site. I have two last notes/recommendations:

1. A few of the examples (the ones I've looked at anyway) require the user to download some data from a Dropbox repository. There isn't necessarily anything wrong with that, but I'd suggest incorporating the required data into the Zenodo repository for this work. Or another Zenodo repository would be fine too. The point would be to have a stable URL that the file reading can be done from, so that the user does not need to download the data and adjust file paths by hand. That could make the examples cleaner and easier for people to run through.

2. In this [Preparation of climate forcing example](https://hydrosolutions.github.io/riversCentralAsia/articles/02-preparation-of-climate-forcing.html), I downloaded the data and adjusted the `gis_path` value for my setup. But I'm still getting this error:
```
> hru_shp <- gen_basinElevationBands(
+     dem_PathN = gis_path, demAggFact = demAggFact, band_interval = band_interval,
+     holeSize_km2 = holeSize_km2, smoothFact = smoothFact)
Error in Ops.units(size_fxn(singles), threshold) :
  both operands of the expression should be "units" objects
```
I'm not sure what to make of it, but thought I'd bring it up in case the authors do!
