### 2022-12-02 a93f5c1
adapt paths in vignettes to integrate zenodo data link
