+++ vignettes/05-snow-calibration.rmd
@@ -20,18 +20,17 @@ knitr::opts_knit$set(
 )
 ```
 
-```{r, message=FALSE, error=FALSE, warning=FALSE}
+```{r, echo=FALSE}
+library(tmap)
+library(sf)
+library(raster)
 library(tidyverse)
 library(lubridate)
-library(sf)
 
 devtools::install_github("hydrosolutions/riversCentralAsia")
 library(riversCentralAsia)
 
-# The Atbashy glacier demo data is available from 
-# https://www.dropbox.com/sh/r0lqggc77ka0uxd/AAChuIyLHHFIfAdgxNKiU2dpa?dl=0
-# Download the data and adapt the path below.  
-data_path <- "../../atbashy_glacier_demo_data/"
+data_path <- "../../riversCentralAsia_ExampleData/"
 ```
 
 This chapter of the vignette shows how to compare simulated snow water equivalent (SWE) to third party data. These code snipets can be used to iteratively adapt the parameters of the snow modules in the HBV models to better represent the SWE of the data product.
@@ -50,7 +49,7 @@ Actually observed SWE is not typically available for hydrological modeling in ma
 
 The downloaded data is subsequently pre-processed to extract the daily average SWE per HRU on the valid pixels of the data set (note that HMASR is not available on areas with permanent snow or ice cover). The following code sniped shows how this can be done.
 
-If you wish to reproduce the code below, please make sure that the [demo data set](https://www.dropbox.com/sh/r0lqggc77ka0uxd/AAChuIyLHHFIfAdgxNKiU2dpa?dl=0){target="_blank"} is downloaded to your computer. 
+If you wish to reproduce the code below, please make sure that the riversCentralAsia_ExampleData (see [README](https://hydrosolutions.github.io/riversCentralAsia/index.html){target="_blank"} is downloaded to your computer. 
 ```{r, eval=FALSE}
 library(tmap)
 library(sf)
@@ -61,12 +60,7 @@ library(lubridate)
 devtools::install_github("hydrosolutions/riversCentralAsia")
 library(riversCentralAsia)
 
-# Download the demo data set from dropbox
-# https://www.dropbox.com/sh/r0lqggc77ka0uxd/AAChuIyLHHFIfAdgxNKiU2dpa?dl=0
-
-# Path to the data directory downloaded from the download link provided above. 
-# Here the data is extracted to a folder called atbashy_glacier_demo_data
-data_path <- "../../atbashy_glacier_demo_data/"
+data_path <- "../../riversCentralAsia_ExampleData/"
 
 # Function to concatenate and mask the HMASR product and extract SWE for each 
 # HRU in the model. 
@@ -140,14 +134,14 @@ extract_HMASR_Atbashy <- function(filespath, year, shape_latlon, varname) {
 }
 
 # Load additional necessary data
-dem <- raster(paste0(data_path, "GIS/16076_DEM.tif"))
-basin <- st_read(paste0(data_path, "GIS/16076_Basin_outline.shp"), 
+dem <- raster(paste0(data_path, "16076_DEM.tif"))
+basin <- st_read(paste0(data_path, "16076_Basin_outline.shp"), 
                  quiet = TRUE)
-shape_latlon <- st_read(paste0(data_path, "GIS/16076_HRU.shp")) |> 
+shape_latlon <- st_read(paste0(data_path, "16076_HRU.shp")) |> 
   st_transform(crs = crs("+proj=longlat +datum=WGS84 +no_defs"))
 
 # Load one example file and display SWE for a random date in the cold season. 
[…]
