### 2022-12-02 #4805 @mabesa
@tonyewong Thanks a lot for your comments!
1. The download of the example data is cumbersome. I have looked into the zenodo option. I ended up uploading the data as a data package and giving instructions in the readme of how the entire package can be downloaded and stored at the same level as the riversCentralAsia package code. I have tested this routine but asked my colleague to re-test when he has time. I'll write here if his tests are successful as well.
2. This is an older bug that should be resolved (I don't get the error). I believe it should disappear on you system when you re-install riversCentralAsia.
