### 2022-12-03 #4805 @tonyewong
Ah, right you are! I stupidly didn't update my package in R but absentmindedly did a `git pull` instead. When I updated the package, it worked perfectly!

With that, I think my review is complete. I recommend accepting the work. Well done!
