### 2022-12-05 hydrosolutions/riversCentralAsia#123 @mengqi-z
Add Examples of Functions for "Preparation for Climate Forcing"
Hi @mabesa,

The instructions in the section [Preparation of climate forcing](https://hydrosolutions.github.io/riversCentralAsia/articles/02-preparation-of-climate-forcing.html) only talk about the `gen_basinElevationBands` function and don't mention much about how `riversCentralAsia` can help prepare climate forcing data, which is a major functionality claimed in the paper. I understand there might be difficulties due to the large size of the CHELSA dataset. Could a simple demo example be added for the usage of the major functions for preparing climate forcing data, such as `biasCorrect_ERA5_CHELSA`, `generate_ERA5_Subbasin_CSV`, etc.? I cannot find the usage of those functions in the book [Modeling of Hydrological Systems in Semi-Arid Central Asia](https://hydrosolutions.github.io/caham_book/) neither.

It would be helpful to mention in this instruction that for advanced user, they can find more details on functions related to climate forcing preparation in this specific section [Downscaling and Bias Correction Using Quantile Mapping](https://hydrosolutions.github.io/caham_book/climate_data.html#sec-bcds-quantile-mapping).

https://github.com/openjournals/joss-reviews/issues/4805
