### 2022-12-05 #4805 @mengqi-z
Hi @mabesa,

I created two new issues about adding examples for few "pre-processing" functions. I understand it might be hard to develop an example if you have to use climate data like CHELSA, which is very large. So, please address those comments accordingly, so that more details are provided on claimed functionalities of the package (e.g., processing of present and future climate forcing) in the paper. Thank you!
