### 2022-12-06 b68ba36
Remove functions that are currently not used in the CAHAM book
