+++ R/biasCorrect_ERA5_CHELSA.R
@@ -1,123 +1,0 @@
-#' Extracts ERA5 basin domain from ERA5 raster bricks and scales fields with
-#' PBCORR CHELSA data so that monthly totals match.
-#' Writes resulting bias corrected (scaled) data bricks to a (newly) created
-#' subdirectory where the raw ERA5 data files are stored.
-#'
-#'
-#' @param dir_ERA5_hourly Directory where hourly ERA5 data annual NetCDF files are stored.
-#' @param dataType_ERA5 Data type, currently only 'tp' or 't2m' available.
-#' @param dir_CHELSA Directory where CHELSA data are stored.
-#' @param startY Starting year of extraction and bias correction (>= 1981)
-#' @param endY Ending year of extraction and bias correction (<= 2013)
-#' @param basinName Name of basin to extract, process and store data
-#' @param basin_shape_LatLon Basin Shapefile in latlon coordinates
-#' @return Dateframe with dates in dd.mm.yyyy hh:mm:ss representation
-#' @family Pre-processing
-#' @export
-biasCorrect_ERA5_CHELSA <- function(dir_ERA5_hourly,dataType_ERA5,dir_CHELSA,startY,endY,basinName,basin_shape_LatLon){
-
-  # Derive area of interest shape
-  aoi_Basin_LatLon <- raster::extent(basin_shape_LatLon)
-  # Function body from above
-  # generate the dates sequence
-  sTime <- base::paste0(startY,'-01-01 01:00:00')
-  eTime <- base::paste0(endY,'-12-31 23:00:00')
-  dateSeq_ERA <- base::seq(base::as.POSIXct(sTime), base::as.POSIXct(eTime), by="hour")
-  dateSeq_ERA <- tibble::tibble(date=dateSeq_ERA, data_bcorr=NA, data_orig=NA)
-  dateSeq_ERA <- dateSeq_ERA %>% dplyr::mutate(month = lubridate::month(date)) %>%
-    dplyr::mutate(year = lubridate::year(date))
-  # Create basin subdirectory (if not already existing) to store dedicated annual files there
-  mainDir <- base::paste0(dir_ERA5_hourly,dataType_ERA5,'/')
-  subDir <- basinName
-  base::ifelse(!base::dir.exists(base::file.path(mainDir, subDir)),
-               base::dir.create(base::file.path(mainDir, subDir)), FALSE)
-
-  # Now starting the time loops
-  for (yr in startY:endY){
-    # progress indicator
-    base::print(base::paste0('PROCESSING YEAR ',yr))
-    # file handling
-    file2Process_ERA <- base::paste0(dataType_ERA5,'/',dataType_ERA5,'_ERA5_hourly_CA_',yr,'.nc')
-    era_data_orig <- raster::brick(paste0(dir_ERA5_hourly,file2Process_ERA))
-    # create date-time tibble with all hours in the corresponding year
-    dateSeq_ERA_year <- dateSeq_ERA %>% dplyr::filter(year == yr)
-
-    # start the month-by-month scaling
-    for (mon in 1:12){
-
-      # sort out files paths as a function of dataType_ERA5
-      if (dataType_ERA5=='t2m'){chlFName='t2m/CHELSA_tmean_'} else {chlFName='pbcorr_tp/CHELSA_tp_bcorr_'}
-      # Load corresponding tmean CHELSA Central Asia File
-      if (mon<10){
-        chelsa_monthly_data <-
-          raster::raster(base::paste0(dir_CHELSA,chlFName,yr,'_0',mon,'_CA_V1.2.1.tif'))
-      } else {
-        chelsa_monthly_data <-
-          raster::raster(base::paste0(dir_CHELSA,chlFName,yr,'_',mon,'_CA_V1.2.1.tif'))
-      }
[…]
