+++ R/generate_ERA5_Subbasin_CSV.R
@@ -1,98 +1,0 @@
-#' Extract subbasin-level and elevation band-specific hourly time series from
-#' ERA5 raster bricks
-#'
-#' Function extracts precipitation (tp) or temperature (t2m) data from raster
-#' bricks and prepares a dataframe for later import in RSMinerve (use
-#' write.table() function to export as csv). If tp and t2m have to be exported,
-#' the function has to be called twice and the resulting tibble columns added.
-#'
-#' @param dir_ERA5_hourly Path to original stored DEM file
-#'   (generate_ERA5_Subbasin_CSV assumes the following folder structure:
-#'   <dir_ERA5_hourly>/<tp or t2m>/<catchmentName>/*.nc)
-#' @param catchmentName Name of catchment for which data should be extracted
-#'   (depends on proper local file storage)
-#' @param dataType Currently, only 'tp' or 't2m' as valid options
-#' @param elBands_shp Subbasin-level shapefile with elevation bands
-#' @param startY Starting year for which data should be made available (assuming
-#'   data 'is' available from the start of that year)
-#' @param endY Ending year from which data should be extracted (assuming data
-#'   'is' actually available until the end of that year)
-#' @return Dataframe tibble with temperature in deg. C. or precipitation in mm/h
-#' @family Pre-processing
-#' @export
-generate_ERA5_Subbasin_CSV <- function(
-    dir_ERA5_hourly,catchmentName,dataType,elBands_shp,startY,endY){
-
-  . <- NULL
-
-  # Generate file list to load
-  yrs <- startY:endY
-  fileList <- tibble::tibble(fName=NA)
-  for (idx in 1:base::length(yrs)){
-    newName <- base::paste0(dataType,'_ERA5_hourly_',catchmentName,'_bcorr_',yrs[idx],'.nc') %>%
-      tibble::as_tibble() %>% dplyr::rename(fName = value)
-    fileList <- fileList %>% tibble::add_row(newName)
-  }
-  fileList <- fileList %>% stats::na.omit()
-  # Since the brick raster data is in +longlat, ensure elBands_shp is in the same crs.
-  elBands_shp_latlon <- sf::st_transform(elBands_shp,crs = sf::st_crs(4326))
-  # Generate date sequence (Note: Dates sequence in accordance with RSMinerve Requirements)
-  dateElBands <- generateSeqDates(startY,endY,'hour') %>% dplyr::slice(-1)
-  # Now, solve that obnoxious time formatting problem for compatibility with RSMinerve (see function posixct2rsminerveChar() for more details)
-  datesChar <- riversCentralAsia::posixct2rsminerveChar(dateElBands$date)
-  datesChar <- datesChar %>% dplyr::rename(Station = value)
-
-  namesElBands <- elBands_shp$name
-  dataElBands_df <- namesElBands %>% purrr::map_dfc(stats::setNames, object = base::list(base::logical())) # fancy trick
-  # to generate an empty dataframe with column names from a vector of characters.
-
-  # Now, loop through the subbasins, one by one.
-  for (yr in 1:length(fileList$fName)){
-    base::print(base::paste0('Processing File: ', fileList$fName[yr]))
-    file2Process_ERA <- fileList$fName[yr]
-    era_data <- raster::brick(base::paste0(dir_ERA5_hourly,dataType,'/',catchmentName,'/',file2Process_ERA))
-    #subbasin_data <- raster::extract(era_data,elBands_shp_latlon) %>% base::lapply(.,colMeans,na.rm=TRUE)
[…]

[…]
