+++ man/biasCorrect_ERA5_CHELSA.Rd
@@ -1,55 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/biasCorrect_ERA5_CHELSA.R
-\name{biasCorrect_ERA5_CHELSA}
-\alias{biasCorrect_ERA5_CHELSA}
-\title{Extracts ERA5 basin domain from ERA5 raster bricks and scales fields with
-PBCORR CHELSA data so that monthly totals match.
-Writes resulting bias corrected (scaled) data bricks to a (newly) created
-subdirectory where the raw ERA5 data files are stored.}
-\usage{
-biasCorrect_ERA5_CHELSA(
-  dir_ERA5_hourly,
-  dataType_ERA5,
-  dir_CHELSA,
-  startY,
-  endY,
-  basinName,
-  basin_shape_LatLon
-)
-}
-\arguments{
-\item{dir_ERA5_hourly}{Directory where hourly ERA5 data annual NetCDF files are stored.}
-
-\item{dataType_ERA5}{Data type, currently only 'tp' or 't2m' available.}
-
-\item{dir_CHELSA}{Directory where CHELSA data are stored.}
-
-\item{startY}{Starting year of extraction and bias correction (>= 1981)}
-
-\item{endY}{Ending year of extraction and bias correction (<= 2013)}
-
-\item{basinName}{Name of basin to extract, process and store data}
-
-\item{basin_shape_LatLon}{Basin Shapefile in latlon coordinates}
-}
-\value{
-Dateframe with dates in dd.mm.yyyy hh:mm:ss representation
-}
-\description{
-Extracts ERA5 basin domain from ERA5 raster bricks and scales fields with
-PBCORR CHELSA data so that monthly totals match.
-Writes resulting bias corrected (scaled) data bricks to a (newly) created
-subdirectory where the raw ERA5 data files are stored.
-}
-\seealso{
-Other Pre-processing: 
-\code{\link{computeDiurnalTemperatureCycle}()},
-\code{\link{convert2HYY}()},
-\code{\link{doQuantileMapping}()},
-\code{\link{gen_HRU_Climate_CSV_RSMinerve}()},
-\code{\link{gen_basinElevationBands}()},
-\code{\link{generate_ERA5_Subbasin_CSV}()},
-\code{\link{loadTabularData}()},
-\code{\link{plotNormDevHYY}()}
-}
-\concept{Pre-processing}
