+++ man/gen_basinElevationBands.Rd
@@ -52,12 +52,9 @@ Note that the DEM should be in UTM coordinates because the code works
 }
 \seealso{
 Other Pre-processing: 
-\code{\link{biasCorrect_ERA5_CHELSA}()},
-\code{\link{computeDiurnalTemperatureCycle}()},
 \code{\link{convert2HYY}()},
 \code{\link{doQuantileMapping}()},
 \code{\link{gen_HRU_Climate_CSV_RSMinerve}()},
-\code{\link{generate_ERA5_Subbasin_CSV}()},
 \code{\link{loadTabularData}()},
 \code{\link{plotNormDevHYY}()}
 }
