+++ vignettes/03-RSMinerveIO.md
@@ -1,0 +1,279 @@
+---
+title: "I/O for RS Minerve"
+output: rmarkdown::html_vignette
+bibliography: [references.bib, HMA_bibliography.bib]
+vignette: >
+  %\VignetteIndexEntry{I/O for RS Minerve}
+  %\VignetteEngine{knitr::knitr}
+  %\VignetteEncoding{UTF-8}
+---
+
+
+
+[RS Minerve](https://crealp.ch/rs-minerve/){target="_blank"} is a free to use hydrologic-hydraulic modelling software. It's accessibility makes is especially suitable for typical Hydromet tasks like flow forecasting. The package riversCentralAsia includes a number of functions to read and write files which interact with RS MINERVE. 
+These tools are interesting for automated writing of forcing files and for reading post-processing and visualization of simulation results or model parameters. 
+
+Using the package `RSMinerveR`, the software can be run from inside R. 
+
+For details on model components and on file formates used in RS MINERVE please refer to the [RS MINERVE manuals](https://crealp.github.io/rsminerve-releases/){target="_blank"}. 
+
+## Reading and writing parameters
+The following RSMinerve objects are currently supported in the `riversCentralAsia` package:\
+- Station\
+- Source\
+- SOCONT\
+- Kinematic\
+- Junction\
+- GSM\
+- HBV92\
+- Comparator
+
+The parameters for these objects are described in the [RS MINERVE User Manual](https://crealp.github.io/rsminerve-releases/){target="_blank"}. 
+
+#### Example
+
+To reproduce the example below, please download the file [Atbashy_PAR_postCal.txt](https://www.dropbox.com/s/y4r44h06jwdnsde/Atbaschy_PAR_postCal.txt?dl=0){target="_blank"} from the demo data folder available from dropbox. Should you want to test a wider number of RS MINERVE I/O functions, we recommend downloading the entire example data folder [here](https://www.dropbox.com/sh/r0lqggc77ka0uxd/AAChuIyLHHFIfAdgxNKiU2dpa?dl=0){target="_blank"}. 
+
+
+```r
+library(tidyverse)
+library(lubridate)
+library(riversCentralAsia)
+
+# Download the file Atbashy_PAR_postCal.txt from https://www.dropbox.com/s/y4r44h06jwdnsde/Atbaschy_PAR_postCal.txt?dl=0
+
+# Adapt the path below 
+parameters_orig <- readRSMParameters("../../atbashy_glacier_demo_data/RSMINERVE/Atbaschy_PAR_postCal.txt") |> 
+  dplyr::filter(Parameters == "Ku [1/d]") 
+
+ggplot(parameters_orig) + 
+  geom_col(aes(Zone, Values), position = "dodge") + 
+  ylab(parameters_orig$Parameters[1]) + 
+  theme_bw()
+```
+
+![Storage coefficient of the upper reservoirs in the HBV models by zone.](figure/unnamed-chunk-2-1.png)
+
+The following shows an example how the model parameters can be changed and written to a RS MINERVE-readable format. 
+
+
+```r
+# Doubble all Ku-values in zone A: 
+parameters <- readRSMParameters(
+  "../../atbashy_glacier_demo_data/RSMINERVE/Atbaschy_PAR_postCal.txt") |> 
+  mutate(Values = ifelse(Parameters == "Ku [1/d]" & Zone == "A", Values * 2, 
+                         Values))
+
+comparison <- parameters_orig |> 
+  mutate(Set = "Original") |> 
+  add_row(parameters |> dplyr::filter(Parameters == "Ku [1/d]") |> 
+            mutate(Set = "Modified"))
+
+ggplot(comparison) + 
[…]
