### 2022-12-06 hydrosolutions/riversCentralAsia#122 @mabesa
Dear @mengqi-z, I am grateful for you vigilance (and sorry to cause a lot of work). I talked to the co-author and we decided to move the function computeDiurnalTemperatureCycle (and the two you mention in issue #123) out of the package because, as you saw, they are currently not used in the CAHAM book.
