+++ vignettes/02-preparation-of-climate-forcing.md
@@ -10,11 +10,12 @@ vignette: >
 
 
 
+## Preparation of a GIS layer with hydrological response unit
 Typically, a river basin is discretized into hydrological response units (HRU) that capture the hydrological properties of a zone within the river basin. Often, HRU are derived from digital elevation models (DEMs) to be able to represent the elevation-dependent timing of snow melt in a basin. This is especially suitable for mountainous basins where gravity flow determines the river runoff. However, in lower lying basins, geology or land use may dominate the hydrology and thus HRU may have to be derived from geological or land use maps.  
 
 The function `gen_basinElevationBands` allows the generation of a shape layer with elevation bands derived from a DEM. Note that you can play with the parameters of the function to adjust the shapes of the elevation bands.   
 
-To reproduce this vignette, please download the DEM [here](https://www.dropbox.com/s/g4cudydy3stcxih/16076_DEM.tif?dl=0). 
+To reproduce this vignette, please make sure you have this package as well as the example data package downloaded (see [README](https://hydrosolutions.github.io/riversCentralAsia/){target="_blank"}). 
 
 
 ```r
@@ -23,7 +24,7 @@ library(lubridate)
 library(timetk)
 library(riversCentralAsia)
 
-gis_path <- "../../atbashy_glacier_demo_data/GIS/16076_DEM.tif"
+gis_path <- "../../riversCentralAsia_ExampleData/16076_DEM.tif"
 crs_project <- "+proj=longlat +datum=WGS84"  # EPSG:4326, latlon WGS84
 
 # Parameter definition for the generation of the elevation bands
@@ -42,9 +43,33 @@ hru_shp %>% plot()
 
 ![Broad elevation bands generated for the Atbashy basin.](figure/unnamed-chunk-2-1.png)
 
-Some minor post-processing is further required and the HRU layer is ready for import to RS Minerve for use in the automated model creation. The open-source book [Modeling of Hydrological Systems in Semi-Arid Central Asia](https://hydrosolutions.github.io/caham_book/){target="_blank"} gives step-by-step instructions for this.   
+Some minor post-processing is further required and the HRU layer is ready for import to RS Minerve for use in the automated model creation. The open-source book Modeling of Hydrological Systems in Semi-Arid Central Asia, [section on geospatial data preparation](https://hydrosolutions.github.io/caham_book/geospatial_data.html){target="_blank"} gives step-by-step instructions for this.   
 
-We recommend the use of daily precipitation and temperature data from the CHELSA data set (made available by WSL)[@karger_climatologies_2017; @karger_high-resolution_2020].  
+## Extraction of climate data on hydrological response units
+We recommend the use of daily precipitation and temperature data from the CHELSA data set (made available by WSL)[@karger_climatologies_2017; @karger_high-resolution_2020]. The data is downloaded from the CHELSA server, then extracted to each hydrological response unit and reformated to the RS MINERVE forcing input format using function ```gen_HRU_Climate_CSV_RSMinerve```. Please note that the global CHELSA data set at 1km grid resolution requires several GB of free storage space on your computer.
+
+This process is demonstrated in detail in the book Modelling of Hydrological Systems in Semi-Arid Central Asia, [section Climate Data - Downscaling and Bias Correction](https://hydrosolutions.github.io/caham_book/climate_data.html#sec-bcds-quantile-mapping){target="_blank"}. Below we show an example useage from this book chapter. This example is reproducible with the (larger) example data set from the book but not with the example data set in this package. 
+
+
+```r
+# Temperature data processing
+temp_or_precip <- "Temperature"
[…]
