### 2022-12-08 654d2fd
update examples using data from zenodo
