### 2022-12-08 b64fe4c
Update vignette, adding examples for climate data preparation
