+++ vignettes/02-preparation-of-climate-forcing.Rmd
@@ -49,7 +49,7 @@ hru_shp <- gen_basinElevationBands(
 hru_shp %>% plot()
 ```
 
-Some minor post-processing is further required and the HRU layer is ready for import to RS Minerve for use in the automated model creation. The open-source book Modeling of Hydrological Systems in Semi-Arid Central Asia, [section on geospatial data preparation](https://hydrosolutions.github.io/caham_book/geospatial_data.html){target="_blank"} gives step-by-step instructions for this.   
+Now, only minor post-processing (like adding a name attribute) is further required and the HRU layer is ready for import to RS Minerve for use in the automated model creation. The open-source book Modeling of Hydrological Systems in Semi-Arid Central Asia, [section on geospatial data preparation](https://hydrosolutions.github.io/caham_book/geospatial_data.html){target="_blank"} gives step-by-step instructions for this.   
 
 ## Extraction of climate data on hydrological response units
 We recommend the use of daily precipitation and temperature data from the CHELSA data set (made available by WSL)[@karger_climatologies_2017; @karger_high-resolution_2020]. The data is downloaded from the CHELSA server, then extracted to each hydrological response unit and reformated to the RS MINERVE forcing input format using function ```gen_HRU_Climate_CSV_RSMinerve```. Please note that the global CHELSA data set at 1km grid resolution requires several GB of free storage space on your computer.
@@ -58,23 +58,110 @@ This process is demonstrated in detail in the book Modelling of Hydrological Sys
 
 ```{r, eval=FALSE}
 # Temperature data processing
-temp_or_precip <- "Temperature"
 hist_obs_tas <- riversCentralAsia::gen_HRU_Climate_CSV_RSMinerve(
   climate_files = <List of CHELSA temperature files>,
   catchmentName = <river_name>,
-  temp_or_precip = <'Temperature' or 'Precipitation'>,
+  temp_or_precip = "Temperature",  # To process precipitation data, type "Precipitation"
   elBands_shp = <hru_shp>,
   startY = <start year>,
   endY = <end year>,
-  obs_freqency = <'hour', 'day', or 'month'>,  # Note that CHELSA data is available at daily or monthly frequency, not hourly.
+  obs_frequency = <'hour', 'day', or 'month'>,  # Note that CHELSA data is available at daily or monthly frequency, not hourly.
   climate_data_type = <'hist_obs', 'hist_sim', or 'fut_sim'>,
-  crs_project = <a projection code, for example '+proj=longlat +datum=WGS84'>)
+  crs_in_use = <a projection code, for example '+proj=longlat +datum=WGS84'>)
 ```
 
-This produces a tibble with the daily average temperatures for each elevation band in the input elBands_shp layer. The same function can be used for other data types like bias-corrected GCM model results. 
+An example using a single input file is given below. 
+```{r, eval=FALSE}
+# Assign names and elevations to the HRU
+# Here we use demo elevations. Typically you would extract elevations from DEM. 
+#   Z <- exactextractr::exact_extract(dem, hru_shp, 'mean')
+hru_shp <- hru_shp |> 
+  mutate(name = paste0("Atb_", layer), 
+         Z = c(1000, 2000, 3000, 4000))  
 
-## Bias correction
[…]
