### 2022-12-08 hydrosolutions/riversCentralAsia#123 @mabesa
Dear @mengqi-z
My colleague had time for testing and I added 2 examples in the Preparation of climate forcing instructions. I also referred to the chapter in the online book which gives detailed step-by-step instructions of how to use the functions in the workflow.

The two functions biasCorrect_ERA5_CHELSA and generate_ERA5_Subbasin_CSV were obsolete and removed from the package.
