### 2022-12-08 #4805 @mabesa
@mengqi-z
Thank you for the good suggestions to improve the documentation. I edited the documentation accordingly and answered to your issues on github.

My colleague and I independently through the examples in the articles again to make sure they now integrate with the zenodo repository for the example data which does not fit into the package.
