### 2023-01-28 #4805 @crvernon
@mabesa we are almost there!

Your paper is looking good.  Could you please make the modification of all non-web address mentions of GitHub be spelled as "GitHub" instead of "github" in the paper?  After that...

So here is what we have left to do:

- [ ]  Conduct a GitHub release of the current reviewed version of the software you now have on the main and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)

- [ ] Please list the DOI of the archived version here
