### 2023-01-30 a796d20
Update paper.md
Corrected spelling of GitHub
