### 2023-02-02 #4805 @kthyng
@mabesa Hi! here are some comments on your paper:

- the units on page 2: should they be km squared?
- the installation section of the paper seems out of place. Is there a reason this is included in the paper instead of just in the README and docs? There is even a commented line in the installation instructions.
- references: please check these in detail. For example, Gudmundsson has a bad character in the citation.
