+++ paper.bib
@@ -1,13 +1,13 @@
-@Manual{qmap_package,
-    title = {qmap: Statistical transformations for post-processing
-      climate model output},
+@misc{qmap_package,
+    title = {qmap: Statistical transformations for post-processing climate model output [R package version 1.0-4]},
     author = {Lukas Gudmundsson},
     year = {2016},
-    note = {R package version 1.0-4},
+    howpublished = {\url{https://cran.r-project.org/web/packages/qmap/}, accessed 2023-02-02},
   }
 
+
 @article{gudmundsson_technical_2012,
-  title = {Technical {{Note}}: {{Downscaling RCM}} Precipitation to the Station Scale Using Statistical Transformations \&ndash; a Comparison of Methods},
+  title = {Technical {{Note}}: {{Downscaling RCM}} Precipitation to the Station Scale Using Statistical Transformations - a Comparison of Methods},
   shorttitle = {Technical {{Note}}},
   author = {Gudmundsson, L. and Bremnes, J. B. and Haugen, J. E. and {Engen-Skaugen}, T.},
   year = {2012},
@@ -22,18 +22,24 @@
   langid = {english}
 }
 
-@article{egusphere-2022-966,
+@article{marti_comprehensive_2023,
   title = {A Comprehensive Open-Source Course for Teaching Applied Hydrological Modelling in {{Central Asia}}},
-  author = {Marti, B. S. and Zhumabaev, A. and Siegfried, T.},
-  year = {2022},
-  journal = {EGUsphere},
-  volume = {2022},
-  pages = {1--18},
-  doi = {10.5194/egusphere-2022-966}
+  author = {Marti, Beatrice Sabine and Zhumabaev, Aidar and Siegfried, Tobias},
+  year = {2023},
+  month = jan,
+  journal = {Hydrology and Earth System Sciences},
+  volume = {27},
+  number = {1},
+  pages = {319--330},
+  publisher = {{Copernicus GmbH}},
+  issn = {1027-5606},
+  doi = {10.5194/hess-27-319-2023},
+  langid = {english}
 }
 
+
 @book{CAHAM:2022,
-  title = {{Modeling of Hydrological Systems in Semi-Arid Central Asia (\url{https://hydrosolutions.github.io/caham_book/})}},
+  title = {{Modeling of Hydrological Systems in Semi-Arid Central Asia (\url{https://hydrosolutions.github.io/caham_book/}, accessed 2023-02-02)}},
   author = {Siegfried, Tobias and Marti, Beatrice},
   publisher = {hydrosolutions},
   url = {https://hydrosolutions.github.io/caham_book/},
@@ -45,30 +51,37 @@
   title = {RS Minerve (v 2.9.1.0)},
   publisher = {CREALP},
   author = {{Bastien Roquier and colleagues}},
-  url = {https://crealp.ch/rs-minerve/},
+  howpublished = {\url{https://crealp.ch/rs-minerve/}, accessed 2023-02-02},
   year = 2021
 }
+
+@misc{roquier_rs_2022,
+  title = {{{RS Minerve Manuals}}},
+  author = {Roquier, Bastien and Fluixa Sanmartin, Javier and Brauchli, Tristan and Baracchini, Theo and Garcia Hernandez, Javier and Foehn, Alain and Paredes Arquiola, Javier and De Cesare, Giovanni},
+  year = {2022},
+  month = may,
+  publisher = {{CREALP}},
[…]
