+++ paper.md
@@ -28,7 +28,7 @@ affiliations:
 
 The R package *riversCentralAsia* includes a set of tools to facilitate and automate data preparation for hydrological modelling. It thus contributes to more reproducible modelling workflows and makes hydrological modelling more accessible to students and to interested professional modellers. 
 
-The package has been developed within the frame of a master level course on applied hydrological modelling in Central Asia [@egusphere-2022-966] and is extensively used in the open-source book "Modelling of Hydrological Systems in Semi-Arid Central Asia" [@CAHAM:2022]. The workflows are further validated within the Horizon 2020 project HYDRO4U [@Hydro4u:2022].
+The package has been developed within the frame of a master level course on applied hydrological modelling in Central Asia [@marti_comprehensive_2023] and is extensively used in the open-source book "Modelling of Hydrological Systems in Semi-Arid Central Asia" [@CAHAM:2022]. The workflows are further validated within the Horizon 2020 project HYDRO4U [@Hydro4u:2022].
 
 While the package has been developed for the Central Asia region, most of the functions are generic and can be used for modelling projects anywhere in the world.
 
@@ -40,23 +40,23 @@ Data preparation comes before hydrological modelling and is actually one of the 
 
 !["Overview over the modelling workflow whos R-components are supported by the riversCentralAsia package (Image source: Marti et al., 2022). Abbreviations are explained in the text. The workflow relies entirely on free, publicly available data and software."](./man/figures/fig_01_model_chain_bw.png){ width=80% }  
 
-The data preparation step covered by the *riversCentralAsia* package includes the derivation of hydrological response units (HRU) using a basin outline and the [SRTM digital elevation model (DEM)](https://doi.org/10.5067/MEaSUREs/SRTM/SRTMGL1.003){target="_blank"}. The [derivation of the basin outline](https://hydrosolutions.github.io/caham_book/geospatial_data.html#sec-catchment-delineation){target="_blank"} and [processing of geospatial layers for import to RS MINERVE](https://hydrosolutions.github.io/caham_book/geospatial_data.html#sec-preparation-of-rsminerve-input-files){target="_blank"} in QGIS is described in detail in @CAHAM:2022.  
+The data preparation step covered by the *riversCentralAsia* package includes the derivation of hydrological response units (HRU) using a basin outline and the [SRTM digital elevation model (DEM)](https://doi.org/10.5067/MEaSUREs/SRTM/SRTMGL1.003){target="_blank"} [@nasa_jpl_nasa_2013]. The [derivation of the basin outline](https://hydrosolutions.github.io/caham_book/geospatial_data.html#sec-catchment-delineation){target="_blank"} and [processing of geospatial layers for import to the hydrological modelling software RS MINERVE](https://hydrosolutions.github.io/caham_book/geospatial_data.html#sec-preparation-of-rsminerve-input-files){target="_blank"} [@rsm:2021] in QGIS [@qgis_development_team_qgis_2022] is described in detail in @CAHAM:2022.  
 
-Although the High Mountain region of Central Asia is generally perceived as a data scarce region, a number of gridded data products are available that form a fair basis for regional hydrological modelling at seasonal time scales. [CHELSA v2.1](https://chelsa-climate.org/){target="_blank"} is a weather data product at 1 km2 grid resolution. The *riversCentralAsia* package includes the function [*gen_HRU_Climate_CSV_RSMinerve*](https://hydrosolutions.github.io/riversCentralAsia/reference/gen_HRU_Climate_CSV_RSMinerve.html){target="_blank"} which extracts CHELSA precipitation or temperature data on the hydrological response units and returns the data in an RS MINERVE readable format.  
[…]
