### 2023-02-02 #4805 @mabesa
- I changed km2 to km squared
- I removed most of the installation instructions but kept the minimum so potential users can get started
-  I exchanged the reference Marti et al., EGU-Sphere to Marti et al. HESS as the article has been published in the meantime.
- The error in the Gudmundsson reference is fixed
- I edited the @misc() fields in bibtext to include the last accessed date
- I added further references to publications/online data sets where suitable
- I allowed myself to make minor edits in the text (e.g. replaced ERA5 with CHELSA) to make the text consistent with the functions in the package. It was an error that still read ERA5, we removed that functionality during the review process.
