### 2023-02-02 #4805 @kthyng
Oh many you need to shift from http style to `10.1038/sdata.2017.122` style?
