### 2023-02-02 6b19b61
Update paper.bib
changed doi style
