+++ paper.bib
@@ -127,7 +127,7 @@
   volume = {4},
   number = {1},
   pages = {170122},
-  doi = {https://doi.org/10.1038/sdata.2017.122}
+  doi = {10.1038/sdata.2017.122}
 }
 
 @misc{chelsa-climatologies-2021,
@@ -135,7 +135,7 @@
 	 publisher = "EnviDat",
 	 title = "Climatologies at high resolution for the earth’s land surface areas",
 	 author = "Dirk Nikolaus Karger and Olaf Conrad and Jürgen Böhner and Tobias Kawohl and Holger Kreft and Rodrigo Wilber Soria-Auza and Niklaus E. Zimmermann and H. Peter Linder and Michael Kessler",
-	 DOI = "http://dx.doi.org/10.16904/envidat.228",
+	 DOI = "10.16904/envidat.228",
 	 url = "https://www.envidat.ch/dataset/chelsa-climatologies"
 }
 
