### 2023-02-02 #4805 @mabesa
I tried your suggestion with removing the http parts
